% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADaMDataPelican}
\alias{ADaMDataPelican}
\title{Example ADaM data for subject profiles, currently 
extracted from the \code{ADaM} datasets for the
Pelican study ('study/study_ADAM_2018_06JUN_15_v2').}
\format{list with data.frame for each dataset}
\description{
This includes the  adverse event ('ADAE'), medical history ('ADMH'), 
exposure ('ADEX'), subject-level ('ADSL'), and laboratory ('ADLB') data.
}
\author{
Laure Cougnaud
}
\keyword{data}
