% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfileReport.R
\name{addLabelReferenceLines}
\alias{addLabelReferenceLines}
\title{Add label of reference lines to a \code{\link[ggplot2]{ggplot}} object}
\usage{
addLabelReferenceLines(gg, refLinesLabels, refLinesTime,
  refLinesColor = "black")
}
\arguments{
\item{gg}{\code{\link[ggplot2]{ggplot}} object}

\item{refLinesLabels}{character vector with label for the reference line(s)}

\item{refLinesTime}{numeric vector with time(x) coordinates for the reference line(s)}

\item{refLinesColor}{vector with color for the reference line(s)}
}
\value{
\code{\link[gtable]{gtable}} objects,
with additional attributes 'nLinesLabelRefLines' containing the number of
lines for the label for the reference lines
}
\description{
Add label of reference lines to a \code{\link[ggplot2]{ggplot}} object
}
\author{
Laure Cougnaud
}
