% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_addReferenceLinesProfilePlot.R
\name{addReferenceLinesProfilePlot}
\alias{addReferenceLinesProfilePlot}
\title{Add reference lines to a profile plot}
\usage{
addReferenceLinesProfilePlot(gg, subjectVar = "USUBJID", refLines = NULL,
  refLinesData = NULL, refLinesTimeVar = NULL, refLinesLabelVar = NULL,
  refLinesColor = "black", refLinesLinetype = "dotted", timeLim = NULL,
  addLabel = FALSE)
}
\arguments{
\item{gg}{\code{\link[ggplot2]{ggplot2}} object}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{refLines}{(optional) nested list with details for reference line(s).
Each sublist contains:
\itemize{
\item{(required) 'label': }{string with label for the reference line}
\item{(required) 'time': }{time(x) coordinate for the reference line,
'dotted' by default}
\item{(optional) 'color': }{color for the reference line,
'black' by default}
\item{(optional) 'linetype': }{linetype for the reference line}
}}

\item{refLinesData}{data.frame with data from which the reference line(s) should be extracted}

\item{refLinesTimeVar}{string, variable of \code{refLinesData} with time for reference line(s)}

\item{refLinesLabelVar}{string, variable of \code{refLinesData} with label for reference line(s)}

\item{refLinesColor}{vector of length 1 with default color for reference line(s)}

\item{refLinesLinetype}{vector of length 1 with default linetype for reference line(s)}

\item{timeLim}{vector of length 2 with time limits}

\item{addLabel}{logical, if TRUE (FALSE by default) add the label of the reference line(s) at the bottom of the plot}
}
\value{
if \code{addLabel} is TRUE, list with:
'gg': \code{\link[ggplot2]{ggplot2}} and 'ggRefLines': \code{\link[ggplot2]{ggplot2}} with labels,
\code{\link[ggplot2]{ggplot2}} otherwise
}
\description{
Add reference lines to a profile plot
}
\author{
Laure Cougnaud
}
