% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_subjectProfileCombine.R
\name{checkTimeExpand}
\alias{checkTimeExpand}
\title{Check if some of the modules are time expanded,
and extract maximum time expand for each module.}
\usage{
checkTimeExpand(listPlots, timeLim = NULL)
}
\arguments{
\item{listPlots}{list of plots}

\item{timeLim}{time limits}
}
\value{
List of time expand for each module
(named by \code{listPlots})
}
\description{
Check if some of the modules are time expanded,
and extract maximum time expand for each module.
}
\author{
Laure Cougnaud
}
