% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_subjectProfileCombine.R
\name{checkTimeTrans}
\alias{checkTimeTrans}
\title{Check if some of the modules are time transformed,
and if some of the plots to align (with specified \code{timeLim})
don't have/have different alignment options.}
\usage{
checkTimeTrans(listPlots, timeLim = NULL)
}
\arguments{
\item{listPlots}{}

\item{timeLim}{}
}
\value{
List of time transformation for each module
(named by \code{listPlots})
}
\description{
Check if some of the modules are time transformed,
and if some of the plots to align (with specified \code{timeLim})
don't have/have different alignment options.
}
\author{
Laure Cougnaud
}
