% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{checkVar}
\alias{checkVar}
\title{Check if specified variable(s) are present in the data.}
\usage{
checkVar(var, data)
}
\arguments{
\item{var}{Character vector with variable(s)
of interest.}

\item{data}{Data.frame with data.}
}
\value{
No returned value,
an error message is triggered if some
variable(s) are not available in the data.
}
\description{
Check if specified variable(s) are present in the data.
}
\author{
Laure Cougnaud
}
