% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfileReport.R
\name{combineVerticallyGGplot}
\alias{combineVerticallyGGplot}
\title{Combine vertically some plots}
\usage{
combineVerticallyGGplot(listPlots, heights, package = c("egg", "cowplot"))
}
\arguments{
\item{listPlots}{list of \code{\link[ggplot2]{ggplot2}} objects}

\item{heights}{vector with heights}

\item{package}{string with package name used to combine the pots, 'egg' or 'cowplot'}
}
\value{
gtable
}
\description{
Combine vertically some plots
}
\author{
Laure Cougnaud
}
