% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{convert_ggplot_shapes_to_plotly_symbols}
\alias{convert_ggplot_shapes_to_plotly_symbols}
\title{Convert [ggplot] shape to [plotly] symbol}
\usage{
convert_ggplot_shapes_to_plotly_symbols(shapes)
}
\arguments{
\item{shapes}{a numeric or character vector representing [ggplot2] shapes}
}
\value{
a string, the corresponding plotly shape name
}
\description{
Convert [ggplot] shape to [plotly] symbol
}
