% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{convert_ggplot_shapes_to_unicode}
\alias{convert_ggplot_shapes_to_unicode}
\title{Convert [ggplot2] shape to unicode character}
\usage{
convert_ggplot_shapes_to_unicode(shapes)
}
\arguments{
\item{shapes}{a numeric or character vector representing [ggplot2] shapes}
}
\value{
a character vector of unicode symbols
}
\description{
Convert [ggplot2] shape to unicode character
}
