% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{convert_shapes_to_plotly_symbols}
\alias{convert_shapes_to_plotly_symbols}
\title{Convert [ggplot2] shapes and some unicode characters to [plotly] symbols}
\usage{
convert_shapes_to_plotly_symbols(shapes)
}
\arguments{
\item{shapes}{a numeric or character vector representing [ggplot2] shapes}
}
\value{
a character vector, the corresponding [plotly] shape name
}
\description{
Convert [ggplot2] shapes and some unicode characters to [plotly] symbols
}
