% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{convertAesVar}
\alias{convertAesVar}
\title{Convert aesthetic variable to factor, converting the empty values ('') to NA}
\usage{
convertAesVar(data, var)
}
\arguments{
\item{data}{data.frame with data}

\item{var}{variable of \code{data} with aesthetic}
}
\value{
updated factor \code{var} variable
}
\description{
Convert aesthetic variable to factor, converting the empty values ('') to NA
}
\author{
Laure Cougnaud
}
