% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataADAM.R
\name{convertToDateTime}
\alias{convertToDateTime}
\title{Convert character vector to date/time object}
\usage{
convertToDateTime(x, format = c("\%Y-\%m-\%dT\%H:\%M", "\%Y-\%m-\%d"),
  colName = NULL, verbose = TRUE)
}
\arguments{
\item{x}{character vector to convert to date/time}

\item{format}{string with possible format(s) of the input date/time in the ADaM dataset.
If multiple are specified, each format is tested successively, until at least one
element in the input vector is converted with the specified format
(non missing, following the approach described in
the \code{format} parameter of the \code{\link{strptime}} function).
See the 'Details' section of the help of the function,
for more information about this format.}

\item{colName}{string with name of column, used in message (if any).}

\item{verbose}{logical, if TRUE (by default) progress messages are printed during execution}
}
\value{
vector of class \code{\link{POSIXct}}
}
\description{
Convert character vector to date/time object
}
\author{
Laure Cougnaud
}
