% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{countNLines}
\alias{countNLines}
\title{Count number of lines ('\\n' character) per character in a vector}
\usage{
countNLines(x)
}
\arguments{
\item{x}{character vector}
}
\value{
numeric vector with same length than \code{x}
containing number of lines in each element
}
\description{
Count number of lines ('\\n' character) per character in a vector
}
\author{
Laure
}
