% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfileReport.R
\name{createSubjectProfileReport}
\alias{createSubjectProfileReport}
\title{Create subject profile report}
\usage{
createSubjectProfileReport(listPlots,
  timeLim = getXLimSubjectProfilePlots(listPlots), landscape = FALSE,
  outputFile = "subjectProfile.pdf", exportFigures = FALSE)
}
\arguments{
\item{listPlots}{list of plots, as returned by the \code{\link{subjectProfileTextPlot}},
\code{\link{subjectProfileEventPlot}} or \code{\link{subjectProfileIntervalPlot}}}

\item{timeLim}{vector with time range}

\item{landscape}{logical, if TRUE the created report is in landscape format}

\item{outputFile}{string, path to the output report}
}
\value{
no returned value, the report is created at the location
specified by \code{outputFile}
}
\description{
Create subject profile report
}
\author{
Laure Cougnaud
}
