% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyEventPlot.R
\name{eventPlotHoverTemplate}
\alias{eventPlotHoverTemplate}
\title{Create [plotly] tool tip hover template for a event plot}
\usage{
eventPlotHoverTemplate(
  paramVal,
  paramLab,
  timeVal,
  timeLab,
  colorLab = NULL,
  colorVal = NULL,
  shapeLab = NULL,
  shapeVal = NULL,
  add_vars = NULL
)
}
\arguments{
\item{paramLab}{Named character vector, 
with label for the parameter variable(s) (\code{paramVar}).\cr
This is used to set the default title.}

\item{timeVal}{the value of `timeVar` for the tooltip}

\item{timeLab}{String, label for \code{timeVar}.
This is used in the message
indicating missing values for \code{timeVar},
and for the default label of the x-axis.}

\item{colorLab}{String, label for \code{colorVar}.}

\item{colorVal}{optional, the value of `colorVar` for the tooltip}

\item{shapeLab}{String, label for \code{shapeVar}.
Set by default to \code{colorLab} if \code{colorVar}
but not \code{shapeVar} is not specified.}

\item{shapeVal}{optional, the value of `shapeVar` for the tooltip}

\item{add_vars}{see [subjectProfileLinePlot()] argument `plotly_args`}

\item{title}{a string, the tool tip title}

\item{paramValueVal}{the value of `paramValueVar` for the tooltip}

\item{paramValueValUnits}{the value of `paramValueVarUnits` for the tooltip}

\item{lln, uln}{optional numeric values, the lower and upper limit normal
values, respectively}
}
\value{
a string with html formatting
}
\description{
Create [plotly] tool tip hover template for a event plot
}
