% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{filterMissingInData}
\alias{filterMissingInData}
\title{Filter missing records in data in the time and y variables,
with informative message.}
\usage{
filterMissingInData(
  data,
  timeVar,
  timeLab = getLabelVar(timeVar, labelVars = labelVars),
  yVar,
  yLab = yVar,
  labelVars = NULL
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{timeVar}{String, variable of \code{data} with time,
displayed in the x axis.\cr
Records with missing time are not displayed in the plot.}

\item{timeLab}{String, label for \code{timeVar}.
This is used in the message
indicating missing values for \code{timeVar},
and for the default label of the x-axis.}

\item{yVar}{String with variable to display in the y-axis.}

\item{yLab}{Label for the y-variable.}
}
\value{
Update data with filtered records +
message in the console.
}
\description{
Filter missing records in data in the time and y variables,
with informative message.
}
\author{
Laure Cougnaud
}
