% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{filterMissingInVar}
\alias{filterMissingInVar}
\title{Filter missing records in data in the time and y variables,
with informative message.}
\usage{
filterMissingInVar(
  data,
  var,
  varLab = getLabelVar(var, labelVars = labelVars),
  labelVars = NULL
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{String with variable of interest.}

\item{yLab}{Label for the variable.}
}
\value{
Update data with filtered records +
message in the console.
}
\description{
Filter missing records in data in the time and y variables,
with informative message.
}
\author{
Laure Cougnaud
}
