% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{formatAdditionalPlotlyHoverVars}
\alias{formatAdditionalPlotlyHoverVars}
\title{Format list of additional hover template variables for [plotly] chart}
\usage{
formatAdditionalPlotlyHoverVars(data, add_vars, labelVars)
}
\arguments{
\item{data}{a data frame with the plotting data}

\item{add_vars}{a list of column names in `data` to add to the hover
template. If some or all elements are unnamed, labels will be taken from
`labelVars`.}

\item{labelVars}{Named character vector with variable labels 
(names are the variable code)}
}
\value{
a named list
}
\description{
Format list of additional hover template variables for [plotly] chart
}
