% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{formatLongLabel}
\alias{formatLongLabel}
\title{Custom formatting function for labels in plot.}
\usage{
formatLongLabel(x, width = 20)
}
\arguments{
\item{x}{character vector with labels to format}

\item{width}{target maximum size. Note: a word
longer that this width won't be split (see \code{\link{strwrap}}).}
}
\value{
vector with formatted labels
}
\description{
Custom formatting function for labels in plot.
}
\author{
Laure Cougnaud
}
