% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{formatParamVar}
\alias{formatParamVar}
\title{Get variable for parameter name}
\usage{
formatParamVar(data, paramVar = NULL, paramGroupVar = NULL,
  revert = FALSE)
}
\arguments{
\item{data}{data.frame with data}

\item{paramVar}{string, variable of \code{data} with parameter}

\item{paramGroupVar}{(optional) character vector with variable(s) of \code{data} with grouping.
If specified, the parameters will be grouped by this(these) variable(s) in the y-axis.}
}
\value{
vector with re-formatted \code{paramVar}, NULL if empty
}
\description{
Get variable for parameter name
}
\author{
Laure Cougnaud
}
