% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{formatTimeLim}
\alias{formatTimeLim}
\title{Format specified \code{timeLim}.}
\usage{
formatTimeLim(
  data,
  subjectVar = "USUBJID",
  timeStartVar,
  timeEndVar,
  timeLim = NULL
)
}
\arguments{
\item{data}{Data.frame with data}

\item{subjectVar}{String, variable of data (and subset data) with subject ID.}

\item{timeStartVar}{string, variable of \code{data} with start of range}

\item{timeEndVar}{string, variable of \code{data} with end of range}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} 
and maximum \code{timeEndVar} per subject.}
}
\value{
Numeric vector of length 2 or list of such element
for each subject.
}
\description{
In case one of the limits if missing,
the corresponding minimum/maximum across subjects is used.
}
\author{
Laure Cougnaud
}
