% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{generate_extra_plotly_hover_text}
\alias{generate_extra_plotly_hover_text}
\title{Generate extra text to include in plotly hover}
\usage{
generate_extra_plotly_hover_text(plotly_hover_text, dat)
}
\arguments{
\item{plotly_hover_text}{Named vector. Plotly hover will include
this information, in form of `<b>Name</b>: value`}

\item{dat}{Data frame, where column `value` will be pulled to make the text}
}
\description{
Generate extra text to include in plotly hover
}
