% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{generate_plotly_hover_text}
\alias{generate_plotly_hover_text}
\title{Generate text to include in plotly hover}
\usage{
generate_plotly_hover_text(
  dat,
  timeVar,
  paramVar,
  colorVar,
  timeShapeVar,
  plotly_hover_text = NULL
)
}
\arguments{
\item{dat}{A row of dataframe}

\item{timeVar}{A string, that determines time mapping}

\item{paramVar}{A string, that determines param mapping}

\item{colorVar}{A string, that determines color mapping}

\item{timeShapeVar}{A string that determines shape mapping}

\item{plotly_hover_text}{A named vector. If not null, 
`generate_extra_plotly_hover_text` will be called to create extra info.}
}
\description{
Generate text to include in plotly hover
}
