% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getAesScaleManual}
\alias{getAesScaleManual}
\title{Get custom 'scale_[type]_manual' function}
\usage{
getAesScaleManual(lab, palette, type)
}
\arguments{
\item{lab}{label for the scale (title of the legend)}

\item{palette}{named vector with color palette}

\item{type}{string with type of scale, e.g. 'color'}
}
\value{
output of the 'scale_[type]_manual' function
}
\description{
Get custom 'scale_[type]_manual' function
}
\author{
Laure Cougnaud
}
