% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{getColorPalettePatientProfile}
\alias{getColorPalettePatientProfile}
\title{Get a color palette of specified length
for patient profile visualization visualization.}
\usage{
getColorPalettePatientProfile(n = NULL, x = NULL, includeNA = TRUE)
}
\arguments{
\item{n}{Integer of length 1, number of elements in palette.}

\item{x}{Vector with elements used for palette.
If factor, the levels are used, otherwise the unique elements of the vector.
Missing values are automatically removed, excepted if
\code{includeNA} is set to TRUE.}

\item{includeNA}{Logical (TRUE by default), 
should NA elements be retained in the palette in case
\code{x} is specified?}
}
\value{
Character vector with colors,
named with the elements in \code{x} if \code{x} is specified.
}
\description{
Get a color palette of specified length
for patient profile visualization visualization.
}
\examples{
# extract longest palette available
getColorPalettePatientProfile(n = 11)
# extract palette for a vector
getColorPalettePatientProfile(x = paste('treatment', 1:4))
# color-blinded palette
getColorPalettePatientProfile(n = 10)
# no color for missing values:
getColorPalettePatientProfile(x = c(NA_character_, "group1"), includeNA = FALSE)
}
\author{
Laure Cougnaud
}
