% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{getGLPGColorPalettePatientProfile}
\alias{getGLPGColorPalettePatientProfile}
\title{Get a Galapagos color palette of specified length
for patient profile visualization visualization.}
\usage{
getGLPGColorPalettePatientProfile(..., includeNA = TRUE)
}
\arguments{
\item{...}{Additional parameters for the 
\code{\link[glpgStyle]{getGLPGColorPalette}} function.}

\item{includeNA}{Logical (TRUE by default), should NA elements be retained in the palette.}
}
\value{
Vector of Galapagos colors,
named with the elements in \code{x} if \code{x} is specified.
}
\description{
This is only a wrapper above \code{\link[glpgStyle]{getGLPGColorPalette}},
including by default missing values (contrary to the default of the
\code{\link[glpgStyle]{getGLPGColorPalette}} function).
}
\author{
Laure Cougnaud
}
