% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{getGLPGLinetypePalettePatientProfile}
\alias{getGLPGLinetypePalettePatientProfile}
\title{Get a Galapagos linetype palette palette of specified length
for patient profile visualization visualization.
This is only a wrapper above \code{\link[glpgStyle]{getGLPGLinetypePalette}},
including by default missing values (contrary to the default of the
code{\link[glpgStyle]{getGLPGLinetypePalette}} function).}
\usage{
getGLPGLinetypePalettePatientProfile(..., includeNA = TRUE)
}
\arguments{
\item{...}{Additional parameters for the 
\code{\link[glpgUtilityFct]{getGLPGLinetypePalette}} function.}

\item{includeNA}{Logical (TRUE by default), should NA elements be retained in the palette.}
}
\value{
character vector values with linetype
}
\description{
Get a Galapagos linetype palette palette of specified length
for patient profile visualization visualization.
This is only a wrapper above \code{\link[glpgStyle]{getGLPGLinetypePalette}},
including by default missing values (contrary to the default of the
code{\link[glpgStyle]{getGLPGLinetypePalette}} function).
}
\author{
Laure Cougnaud
}
