% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{getGLPGShapePalettePatientProfile}
\alias{getGLPGShapePalettePatientProfile}
\title{Get a Galapagos shape palette palette of specified length
for patient profile visualization visualization.}
\usage{
getGLPGShapePalettePatientProfile(..., includeNA = TRUE, asText = TRUE)
}
\arguments{
\item{...}{Additional parameters for the 
\code{\link[glpgUtilityFct]{getGLPGShapePalette}} function.}

\item{includeNA}{Logical (TRUE by default), should NA elements be retained in the palette.}

\item{asText}{Logical (TRUE by default), should the palette
be expressed as integer (base R plot and ggplot2 compatible)
or in text format (e.g. required if combined with unicode symbols in ggplot2)?}
}
\value{
vector of shape values
}
\description{
This is only a wrapper above \code{\link[glpgUtilityFct]{getGLPGShapePalette}},
including by default missing values and formatting the symbols as text
(in case mixed with Unicode).
}
\author{
Laure Cougnaud
}
