% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getLabelVar}
\alias{getLabelVar}
\title{Get label(s) for a variable of the dataset}
\usage{
getLabelVar(var, data = NULL, labelVars = NULL)
}
\arguments{
\item{var}{variable of data}

\item{data}{data.frame with data}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
string with label, \code{var} is no label is available
}
\description{
The label(s) are extracted either:
\itemize{
\item{if \code{data} is specified: }{from the 'label' attribute 
of the corresponding column in \code{data}
}
\item{if \code{labelVars} is specified: }{
from the specified vector of labels}
}
}
\author{
Laure Cougnaud
}
