% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataADAM.R
\name{getLabelVars}
\alias{getLabelVars}
\title{Get label of the variables in ADaM dataset(s)}
\usage{
getLabelVars(dataList)
}
\arguments{
\item{dataList}{list with data.frame with ADaM dataset(s)}
}
\value{
named vector with variable labels
}
\description{
Get label of the variables in ADaM dataset(s)
}
\author{
Laure Cougnaud
}
