% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{getLinetypePalettePatientProfile}
\alias{getLinetypePalettePatientProfile}
\title{Get a linetype palette palette of specified length
for patient profile visualization visualization.}
\usage{
getLinetypePalettePatientProfile(n = NULL, x = NULL, includeNA = TRUE)
}
\arguments{
\item{n}{Integer of length 1, number of elements in palette.}

\item{x}{Vector with elements used for palette.
If factor, the levels are used, otherwise the unique elements of the vector.
Missing values are automatically removed, excepted if
\code{includeNA} is set to TRUE.}

\item{includeNA}{Logical (TRUE by default), 
should NA elements be retained in the palette in case
\code{x} is specified?}
}
\value{
character vector values with linetype
}
\description{
Note that 7 unique symbols are available at maximum
(replicated if necessary).
}
\examples{
# extract longest linetype palette available
getLinetypePalettePatientProfile(n = 6)
# extract palette for a vector
getLinetypePalettePatientProfile(x = paste('treatment', 1:4))
# include missing
getLinetypePalettePatientProfile(x = c(NA_character_, "group1"), includeNA = TRUE)
}
\author{
Laure Cougnaud
}
