% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getNLinesLabel}
\alias{getNLinesLabel}
\title{Get number of lines for specific label}
\usage{
getNLinesLabel(gg, elName = c("x", "y", "title", "caption"),
  elNLines = NULL)
}
\arguments{
\item{gg}{\code{\link[ggplot2]{ggplot2}} object}

\item{elName}{string with name of label to extract,
among 'x', 'y' and 'title'}

\item{elNLines}{(optional) integer with number of lines,
by default 2 for 'x'/'y' and 3 for 'title'}
}
\value{
vector with (approximated) number of lines
}
\description{
Get number of lines for specific label
}
\author{
Laure Cougnaud
}
