% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getNLinesLegend}
\alias{getNLinesLegend}
\title{Get number of lines in the legend of a \code{\link[ggplot2]{ggplot2}} object}
\usage{
getNLinesLegend(gg)
}
\arguments{
\item{gg}{\code{\link[ggplot2]{ggplot2}} object}
}
\value{
integer with (approximated) number of lines
}
\description{
Get number of lines in the legend of a \code{\link[ggplot2]{ggplot2}} object
}
\author{
Laure Cougnaud
}
