% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getPathTemplate}
\alias{getPathTemplate}
\title{Get path of the report template 
in the \code{patientProfilesVis} package}
\usage{
getPathTemplate(file)
}
\arguments{
\item{file}{file name (with extension)}
}
\value{
String with path to the template in the 
installed \code{patientProfilesVis} package
}
\description{
Get path of the report template 
in the \code{patientProfilesVis} package
}
\author{
Laure Cougnaud
}
