% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_combineVerticallyGGplot.R
\name{getSplitVectorByInt}
\alias{getSplitVectorByInt}
\title{Split/combine a vector of size(s) to have a fixed combined size}
\usage{
getSplitVectorByInt(sizes, max = NULL)
}
\arguments{
\item{sizes}{vector with size}

\item{max}{integer with maximum combined size in output,
Inf by default.}
}
\value{
vector of same length as \code{sizeVect},
containing corresponding class
}
\description{
Split/combine a vector of size(s) to have a fixed combined size
}
\author{
Laure Cougnaud
}
