% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getWidthPlot}
\alias{getWidthPlot}
\title{Get width for a plot for a certain page layout}
\usage{
getWidthPlot(margin = subjectProfileReportFormat()$margin,
  landscape = subjectProfileReportFormat()$landscape,
  aspectRatio = subjectProfileReportFormat()$aspectRatio)
}
\arguments{
\item{margin}{margin in inches}

\item{landscape}{logical, if TRUE the created report is in landscape format}

\item{aspectRatio}{ratio between size of image in inches 
(derived from specified margin, landscape and heightLineIn)
and real size for exported image}
}
\value{
width for the plot in inches
}
\description{
Get width for a plot for a certain page layout
}
\author{
Laure Cougnaud
}
