% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfileReport.R
\name{getXLimSubjectProfilePlots}
\alias{getXLimSubjectProfilePlots}
\title{Get limits for a list of plots.}
\usage{
getXLimSubjectProfilePlots(listPlots)
}
\arguments{
\item{listPlots}{list of \code{subjectProfile[X]Plot} plots}
}
\value{
vector of length 2 with limits for the x-axis
}
\description{
These limits are extracted from the maximal range
of the x-coordinates across all plots
}
\author{
Laure Cougnaud
}
