% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{isSubjectProfileTimeVariant}
\alias{isSubjectProfileTimeVariant}
\title{Check if the all profile(s) is/are 'time-variant',
so not a subject profile 'text' module or empty plot}
\usage{
isSubjectProfileTimeVariant(gg, empty = TRUE)
}
\arguments{
\item{gg}{object of class \code{subjectProfileX} 
(and \code{\link[ggplot2]{ggplot}}) or potentially
nested list of such objects.}

\item{empty}{Logical, should empty subject profile be
considered as time-variant?}
}
\value{
Logical, is plot time variant?
}
\description{
Check if the all profile(s) is/are 'time-variant',
so not a subject profile 'text' module or empty plot
}
\author{
Laure Cougnaud
}
