% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitPrintListPlotsKnitr.R
\name{knitPrintListPlotsKnitr}
\alias{knitPrintListPlotsKnitr}
\title{include list of plots with different figure height, size, label in a knitr document}
\usage{
knitPrintListPlotsKnitr(plotsList, generalLabel = "plotsList",
  labels = paste0(generalLabel, seq_along(plotsList)), type = c("ggplot2",
  "plotly"), includeNewpage = TRUE, ...)
}
\arguments{
\item{plotsList}{list of plots, e.g. \code{ggplot} objects
from the \code{ggplot2} package or from the \code{plotly} packags}

\item{generalLabel}{general label for the chunks, used to build the \code{labels}.
The labels are constructed as '\code{generalLabel}[i]',
with i the plot number (from sequence spanning the length of \code{plotsList}).
Only use if \code{labels} is not specified.}

\item{labels}{vector with labels for each chunk}

\item{type}{string with plot type, 'ggplot2' or 'plotly'}

\item{includeNewpage}{logical, if TRUE include newpage after each plot}

\item{...}{any chunk parameters, will be replicated if necessary
See \code{\link{knitr}[opts_chunk]} for further details on available options.}
}
\value{
no returned value, a text is printed with chunk content
}
\description{
Each plot will be included in a separated chunk.
Reason: fig.height and fig.width options are not vectorized.
Note: the 'results' knitr options of the chunk calling this function
should be set to 'asis'.
Note that a (one-level) list of plotly plots can be included directly via
htmltools::tagList(listPlots).
}
\author{
Laure Cougnaud
}
