% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyLinePlot.R
\name{linePlotHoverTemplate}
\alias{linePlotHoverTemplate}
\title{Create [plotly] tool tip hover template for a line plot}
\usage{
linePlotHoverTemplate(
  data,
  paramFacetVar,
  paramValueVar,
  paramValueVarUnits = NULL,
  timeLab,
  timeVar,
  colorLab = NULL,
  colorVar = NULL,
  shapeLab = NULL,
  shapeVar = NULL,
  paramValueRangeVar = NULL,
  add_vars = NULL,
  labelVars
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{paramValueVar}{String, variable of \code{data} with parameter value to
represent.\cr Records with missing values are discarded.}

\item{timeLab}{String, label for \code{timeVar}.
This is used in the message
indicating missing values for \code{timeVar},
and for the default label of the x-axis.}

\item{timeVar}{String, variable of \code{data} with time,
displayed in the x axis.\cr
Records with missing time are not displayed in the plot.}

\item{colorLab}{String, label for \code{colorVar}.}

\item{colorVar}{String, variable of \code{data} with color.}

\item{shapeLab}{String, label for \code{shapeVar}.
Set by default to \code{colorLab} if \code{colorVar}
but not \code{shapeVar} is not specified.}

\item{shapeVar}{String, variable of \code{data} for shape of the points.
By default, same as \code{colorVar}.}

\item{paramValueRangeVar}{Character vector of length 2 containing variables
of \code{data} with minimum and maximum value for \code{paramValueVar},
typically reference range indicators.\cr Range can differ per parameter and
even per time point. This range is represented as a ribbon in the plot
background. e.g. to represent the reference range of the variable.}

\item{add_vars}{see [subjectProfileLinePlot()] argument `plotly_add_vars`}

\item{labelVars}{Named character vector with variable labels 
(names are the variable code)}
}
\value{
a character vector of html formatted strings
}
\description{
Create [plotly] tool tip hover template for a line plot
}
