% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeTransformations.R
\name{logPlotlyXAxis}
\alias{logPlotlyXAxis}
\title{Log x-axis variables in the data}
\usage{
logPlotlyXAxis(data, xvars, log_x_axis)
}
\arguments{
\item{data}{a data frame}

\item{xvars}{a character vector, the column names in `data` that contain the
x-axis values}

\item{log_x_axis}{a string, indicating how the x-axis variable should be
scaled. If `'neg'`, only the negative values are scaled, if `'pos'` only
the positive values are scaled, if `'both'` the positive and negative
values are scaled.}
}
\value{
a named list: \itemize{
  \item{`data`}{The modified data frame with new variable(s)}
  \item{`footnote`}{A string, a footnote that can be added to the caption 
    that describes how the x-axis was logged and which states that the 
    tooltip x-axis values is not the logged values but rather the original 
    value}
}
}
\description{
Provided column name(s) in a data frame, this function creates a logged 
version of those column(s) in the same data frame. The new variables in the 
data frame are named the same as the old one with `"Log"` appended to the 
end of their names. You can apply the log function to the entire x-axis, only
the negative values, or only the positive values. Logging the negative values 
is often useful for plots of medical history where a few items have extreme 
negative x-axis values which results in compression of the positive x-axis 
values, when the positive x-axis values may be of more interest.
}
