% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientProfilesVis-common-args-doc.R
\name{patientProfilesVis-common-args}
\alias{patientProfilesVis-common-args}
\title{Arguments used across the functions of the patientProfilesVis package.}
\arguments{
\item{data}{Data.frame with data.}

\item{colorVar}{String, variable of \code{data} with color.}

\item{colorLab}{String, label for \code{colorVar}.}

\item{colorPalette}{Named vector with color palette.
The variable should be named with the corresponding element
in \code{colorVar}.}

\item{xLab}{String, label for the x-axis.}

\item{yLab}{String, label for the y-axis.}

\item{label}{String, label for the visualization.
This label is used when the visualization is 
included in the final patient profile report,
in case this visualization is empty (no data available)
for a specific subject.}

\item{paramVar}{Character vector with variable(s) of \code{data} 
with parameters. Variable content is displayed in the y-axis.}

\item{paramLab}{Named character vector, 
with label for the parameter variable(s) (\code{paramVar}).\cr
This is used to set the default title.}
}
\description{
Arguments used across the functions of the patientProfilesVis package.
}
