% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{patientProfilesVis-palette}
\alias{patientProfilesVis-palette}
\title{Parameters for all patient profiles visualization palette functions.}
\arguments{
\item{includeNA}{Logical (TRUE by default), 
should NA elements be retained in the palette in case
\code{x} is specified?}
}
\description{
Parameters for all patient profiles visualization palette functions.
}
