% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyEventPlot.R
\name{plotlyEventPlot}
\alias{plotlyEventPlot}
\title{Create a [plotly] event plot for a subject}
\usage{
plotlyEventPlot(
  data,
  paramLab,
  colorVar = NULL,
  colorLab = NULL,
  colorPalette = getColorPalettePatientProfile(n = 1),
  shapeVar = NULL,
  shapeLab = NULL,
  shapePalette = NULL,
  alpha = 1,
  timeVar,
  timeLab,
  timeLim = NULL,
  xLab,
  yLab,
  title,
  labelVars = NULL,
  margin = list(l = 50, r = 50, b = 50, t = 50, pad = 4),
  log_footnote_y_shift = -0.12,
  log_x_axis = NULL,
  add_vars = NULL
)
}
\arguments{
\item{data}{a data frame with data for 1 subject only}

\item{paramLab}{Named character vector, 
with label for the parameter variable(s) (\code{paramVar}).\cr
This is used to set the default title.}

\item{colorVar}{String, variable of \code{data} with color.}

\item{colorLab}{String, label for \code{colorVar}.}

\item{colorPalette}{Named vector with color palette.
The variable should be named with the corresponding element
in \code{colorVar}.\cr
Colors can also be defined for the entire session, by setting
\code{options(patientProfilesVis.colors = X)} with X either:
\itemize{
\item{a vector with colors}
\item{a function returning a vector of colors for a specified number
of elements (\code{viridis} by default)}
}}

\item{shapeVar}{String, variable of \code{data} for shape of the points.
By default, same as \code{colorVar}.}

\item{shapeLab}{String, label for \code{shapeVar}.
Set by default to \code{colorLab} if \code{colorVar}
but not \code{shapeVar} is not specified.}

\item{shapePalette}{Named character vector with shape palette
for \code{shapeVar}.
The variable should be named with the corresponding element
in \code{shapeVar}.\cr
Shapes can also be defined for the entire session, by setting
\code{options(patientProfilesVis.shapes = X)} with X either:
\itemize{
\item{a vector with shapes}
\item{a function returning a vector of shapes for a specified number
of elements}
}
Note it is advised to specify the shapes as character,
e.g. 'cross' instead of \code{4}, in case Unicode
symbols should also be used.}

\item{alpha}{Numeric with transparency, 1 by default.}

\item{timeVar}{String, variable of \code{data} with time,
displayed in the x axis.\cr
Records with missing time are not displayed in the plot.}

\item{timeLab}{String, label for \code{timeVar}.
This is used in the message
indicating missing values for \code{timeVar},
and for the default label of the x-axis.}

\item{timeLim}{(optional) Vector of length 2 with time limits (x-axis).
If not specified, these are extracted from the minimum \code{timeStartVar} 
and maximum \code{timeEndVar} per subject.\cr
The time limits are stored as attributes of the plots,
used to align the plots in the final report.}

\item{xLab}{String, label for the x-axis.}

\item{yLab}{String, label for the y-axis.}

\item{title}{String, title for the plot.}

\item{labelVars}{Named character vector with variable labels 
(names are the variable code)}

\item{margin}{see [subjectProfileEventPlot()] argument `plotly_args`}

\item{log_footnote_y_shift}{see [subjectProfileEventPlot()] argument
`plotly_args`}

\item{log_x_axis}{see [subjectProfileEventPlot()] argument `plotly_args`}

\item{add_vars}{see [subjectProfileEventPlot()] argument `plotly_args`}
}
\value{
a [plotly] object
}
\description{
Create a [plotly] event plot for a subject
}
\details{
This function is designed to be called by [subjectProfileEventPlot()].
}
