% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyIntervalPlot.R
\name{plotlyIntervalPlot}
\alias{plotlyIntervalPlot}
\title{Create a [plotly] interval plot for a subject}
\usage{
plotlyIntervalPlot(
  data,
  paramVar,
  paramLab,
  timeStartVar,
  timeStartLab,
  timeEndVar,
  timeEndLab,
  timeLim = NULL,
  colorVar = NULL,
  colorLab = NULL,
  colorPalette = getColorPalettePatientProfile(n = 1),
  alpha = 1,
  timeStartShapeVar,
  timeEndShapeVar,
  shapeLab,
  shapePalette = getShapePalettePatientProfile(n = 1),
  shapeSize,
  title,
  xLab,
  yLab,
  caption = NULL,
  labelVars = NULL,
  margin = list(l = 50, r = 50, b = 100, t = 50, pad = 4),
  caption_y_shift = -0.12,
  log_x_axis = NULL,
  add_vars = NULL
)
}
\arguments{
\item{data}{a data frame with data for 1 subject only}

\item{paramVar}{Character vector with variable(s) of \code{data} 
with parameters. Variable content is displayed in the y-axis.}

\item{paramLab}{Named character vector, 
with label for the parameter variable(s) (\code{paramVar}).\cr
This is used to set the default title.}

\item{timeStartVar}{String, variable of \code{data} 
with start of time interval.}

\item{timeStartLab}{String, label for \code{timeStartVar},
displayed in a message and in the plot caption.}

\item{timeEndVar}{String, variable of \code{data} 
with end of time interval.}

\item{timeEndLab}{String, label for \code{timeEndVar},
displayed in a message and in the plot caption.}

\item{timeLim}{(optional) Vector of length 2 with time limits (x-axis).
If not specified, these are extracted from the minimum \code{timeStartVar} 
and maximum \code{timeEndVar} per subject.\cr
The time limits are stored as attributes of the plots,
used to align the plots in the final report.}

\item{colorVar}{String, variable of \code{data} with color.}

\item{colorLab}{String, label for \code{colorVar}.}

\item{colorPalette}{Named vector with color palette.
The variable should be named with the corresponding element
in \code{colorVar}.\cr
Colors can also be defined for the entire session, by setting
\code{options(patientProfilesVis.colors = X)} with X either:
\itemize{
\item{a vector with colors}
\item{a function returning a vector of colors for a specified number
of elements (\code{viridis} by default)}
}}

\item{alpha}{Numeric with transparency, 1 by default.}

\item{timeStartShapeVar}{(optional) String, variable of \code{data} 
used for the shape of the symbol displayed 
at the start of the time interval.\cr
If not specified, default shape palette is used,
see section 'Time interval representation'.}

\item{timeEndShapeVar}{String, variable of \code{data} 
used for the shape of the symbol 
displayed at the end of the time interval.
If not specified, default shape palette is used,
see section 'Time interval representation'.}

\item{shapeLab}{String, label for \code{shapeVar}.
Set by default to \code{colorLab} if \code{colorVar}
but not \code{shapeVar} is not specified.}

\item{shapePalette}{Named character vector with shape palette
for \code{shapeVar}.
The variable should be named with the corresponding element
in \code{shapeVar}.\cr
Shapes can also be defined for the entire session, by setting
\code{options(patientProfilesVis.shapes = X)} with X either:
\itemize{
\item{a vector with shapes}
\item{a function returning a vector of shapes for a specified number
of elements}
}
Note it is advised to specify the shapes as character,
e.g. 'cross' instead of \code{4}, in case Unicode
symbols should also be used.}

\item{shapeSize}{Size for symbols (only used if \code{timeStartShapeVar}/\code{timeEndShapeVar} is specified).}

\item{title}{String, title for the plot.}

\item{xLab}{String, label for the x-axis.}

\item{yLab}{String, label for the y-axis.}

\item{caption}{(optional) String with caption (NULL for no caption). 
By default the caption contains information on the imputation strategy 
for missing time.}

\item{labelVars}{Named character vector with variable labels 
(names are the variable code)}

\item{margin}{see [subjectProfileIntervalPlot()] argument `plotly_args`}

\item{caption_y_shift}{see [subjectProfileIntervalPlot()] argument
`plotly_caption_y_shift`}

\item{log_x_axis}{see [subjectProfileIntervalPlot()] argument
`plotly_log_x_axis`}

\item{add_vars}{see [subjectProfileIntervalPlot()] argument `plotly_add_vars`}
}
\value{
a [plotly] object
}
\description{
Create a [plotly] interval plot for a subject
}
\details{
This function is designed to be called by [subjectProfileIntervalPlot()].
}
