% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_subjectProfileCombine.R
\name{prepareSubjectProfile}
\alias{prepareSubjectProfile}
\title{prepare list of subject profile (s) to be combined with the \code{\link{combineVerticallyGGplot}}}
\usage{
prepareSubjectProfile(..., labels, timeLim = NULL, refLines = NULL,
  refLinesData = NULL, refLinesTimeVar = NULL,
  refLinesLabelVar = NULL, subjectVar = "USUBJID")
}
\arguments{
\item{...}{list of subject profiles (across modules)}

\item{labels}{string with labels for the plots}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} 
and maximum \code{timeEndVar} per subject.}

\item{refLines}{(optional) nested list with details for reference line(s).
Each sublist contains:
\itemize{
\item{(required) 'label': }{string with label for the reference line}
\item{(required) 'time': }{time(x) coordinate for the reference line,
'dotted' by default}
\item{(optional) 'color': }{color for the reference line,
'black' by default}
\item{(optional) 'linetype': }{linetype for the reference line}
}}

\item{refLinesData}{data.frame with data from which the reference line(s) should be extracted}

\item{refLinesTimeVar}{string, variable of \code{refLinesData} with time for reference line(s)}

\item{refLinesLabelVar}{string, variable of \code{refLinesData} with label for reference line(s)}

\item{subjectVar}{string, variable of \code{data} with subject ID}
}
\value{
\code{subjectProfilePlot} object, containing the combined
profile plots
}
\description{
prepare list of subject profile (s) to be combined with the \code{\link{combineVerticallyGGplot}}
}
\author{
Laure Cougnaud
}
