% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfileReport.R
\name{print.subjectProfilePlot}
\alias{print.subjectProfilePlot}
\title{convenience function to print a subject profile plot (of class \code{gtable}).
This plot is printed with the \code{grid.draw} function}
\usage{
\method{print}{subjectProfilePlot}(x, ...)
}
\arguments{
\item{x}{object of class \code{subjectProfilePlot}}

\item{...}{additional arguments, currently not used}
}
\value{
no returned value, a plot is drawn in the current window
}
\description{
convenience function to print a subject profile plot (of class \code{gtable}).
This plot is printed with the \code{grid.draw} function
}
\author{
Laure Cougnaud
}
