% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdtmDataPelican}
\alias{sdtmDataPelican}
\title{Example data for subject profiles, currently 
extracted from the \code{sdtm} datasets for the
Pelican study ('GLPG2737-CL-202-UNBLINDED/CRG10GLP01_SDTM_2018_06JUN_15_v2').}
\format{list with data.frame for each dataset}
\description{
This includes the  adverse event ('ae'), medical history ('mh'), 
demographics ('dm'), and laboratory ('lb') data.
}
\author{
Laure Cougnaud
}
\keyword{data}
