% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfileReport.R
\name{subjectProfileCombineOnce}
\alias{subjectProfileCombineOnce}
\title{combine multiple \code{\link[ggplot2]{ggplot2}} objects}
\usage{
subjectProfileCombineOnce(..., labels, timeLim = NULL, maxNLines = NULL,
  refLines = NULL, refLinesData = NULL, refLinesTimeVar = NULL,
  refLinesLabelVar = NULL, subjectVar = "USUBJID")
}
\arguments{
\item{...}{\code{\link[ggplot2]{ggplot2}} objects}

\item{labels}{string with labels for the plots}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} and maximum \code{timeEndVar}}

\item{maxNLines}{maximum number of lines for a combined plot,
to fit in the page height}

\item{refLines}{(optional) nested list with details for reference line(s).
Each sublist contains:
\itemize{
\item{(required) 'label': }{string with label for the reference line}
\item{(required) 'time': }{time(x) coordinate for the reference line,
'dotted' by default}
\item{(optional) 'color': }{color for the reference line,
'black' by default}
\item{(optional) 'linetype': }{linetype for the reference line}
}}

\item{refLinesData}{data.frame with data from which the reference line(s) should be extracted}

\item{refLinesTimeVar}{string, variable of \code{refLinesData} with time for reference line(s)}

\item{refLinesLabelVar}{string, variable of \code{refLinesData} with label for reference line(s)}

\item{subjectVar}{string, variable of \code{data} with subject ID}
}
\value{
\code{subjectProfilePlot} object, containing the combined
profile plots
}
\description{
combine multiple \code{\link[ggplot2]{ggplot2}} objects
}
\author{
Laure Cougnaud
}
