% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileIntervalPlot.R
\name{subjectProfileIntervalPlot}
\alias{subjectProfileIntervalPlot}
\title{Create plot of subject profiles with segments for range of parameters.}
\usage{
subjectProfileIntervalPlot(
  data,
  paramVar,
  paramVarSep = " - ",
  paramLab = toString(getLabelVar(paramVar, labelVars = labelVars)),
  paramGroupVar = NULL,
  timeStartVar,
  timeStartLab = getLabelVar(timeStartVar, labelVars = labelVars),
  timeEndVar,
  timeEndLab = getLabelVar(timeEndVar, labelVars = labelVars),
  timeLabel = "time",
  subjectVar = "USUBJID",
  subjectSubset = NULL,
  subsetData = NULL,
  subsetVar = NULL,
  subsetValue = NULL,
  timeImpType = c("minimal", "data-by-subject", "none"),
  timeLim = NULL,
  timeLimData = NULL,
  timeLimStartVar = NULL,
  timeLimStartLab = getLabelVar(timeLimStartVar, labelVars = labelVars),
  timeLimEndVar = NULL,
  timeLimEndLab = getLabelVar(timeLimEndVar, labelVars = labelVars),
  timeTrans = NULL,
  timeExpand = NULL,
  timeAlign = TRUE,
  rangeSimilarStartEnd = NULL,
  xLab = paste(c(timeStartLab, timeEndLab), collapse = "/"),
  yLab = "",
  colorVar = NULL,
  colorLab = getLabelVar(colorVar, labelVars = labelVars),
  colorPalette = NULL,
  alpha = 1,
  timeStartShapeVar = NULL,
  timeEndShapeVar = NULL,
  shapePalette = NULL,
  shapeLab = toString(unique(getLabelVar(c(timeStartShapeVar, timeEndShapeVar),
    labelVars = labelVars))),
  shapeSize = rel(3),
  title = paramLab,
  label = title,
  labelVars = NULL,
  formatReport = subjectProfileReportFormat(),
  paging = TRUE
)
}
\arguments{
\item{data}{data.frame with data}

\item{paramVar}{string, variable of \code{data} with parameter (used in the y-axis)}

\item{paramVarSep}{string with character(s) used to concatenate multiple 
\code{paramVar}, ' - ' by default.}

\item{paramLab}{string, label for \code{paramVar}}

\item{paramGroupVar}{(optional) character vector with variable(s) of \code{data} with grouping.
If specified, the parameters will be grouped by this(these) variable(s) in the y-axis.}

\item{timeStartVar}{string, variable of \code{data} with start of range}

\item{timeStartLab}{String, label for \code{timeStartVar}.}

\item{timeEndVar}{string, variable of \code{data} with end of range}

\item{timeEndLab}{String, label for \code{timeEndVar}.}

\item{timeLabel}{string with general time label, used
in the footnote for the explanation of the arrow, 'time' by default.}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{subjectSubset}{(optional) Character vector with subjects of interest 
(available in \code{subjectVar} of \code{data}), NULL by default.}

\item{subsetData}{Data.frame from which subset based on \code{subjectVar}/\code{subsetValue}
should be extracted. If not specified, \code{data} is used.
Records from \code{data} are then mapped by the \code{subjectVar} variable.}

\item{subsetVar}{variable of \code{data} used for subsetting}

\item{subsetValue}{character vector with value(s) of interest to consider for 
\code{subsetVar}}

\item{timeImpType}{String with imputation type: 'minimal' (default),
'data-based' or 'none'.}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} 
and maximum \code{timeEndVar} per subject.}

\item{timeLimData}{data.frame with data used to extract time limits per subject}

\item{timeLimStartVar}{string, variable of \code{timeLimData} with time start}

\item{timeLimEndVar}{string, variable of \code{timeLimData} with time end}

\item{timeTrans}{ggplot2 transformation (see \code{\link[scales]{trans_new}}),
e.g. produced by the \code{\link{getTimeTrans}} function.}

\item{timeExpand}{Vector of range expansion constants for the time axis
(see \code{expand} parameter of the \code{\link[ggplot2]{scale_x_continuous}} function).}

\item{timeAlign}{Logical, if TRUE (by default)
the different plots are horizontally aligned.
If set to FALSE, each plot has its own time-limits.}

\item{rangeSimilarStartEnd}{numeric, if a record has the same
\code{timeStartVar} and \code{timeEndVar}, what should be the range of the segment?
By default, a thousandth of the range of \code{timeLim}
(if not specified the time range available in the data).}

\item{xLab}{string, label for the x-axis}

\item{yLab}{string, label for the y-axis}

\item{colorVar}{string, variable of \code{data} with color.}

\item{colorLab}{string, label for \code{colorVar}}

\item{colorPalette}{named vector with color for \code{colorVar}}

\item{alpha}{Numeric with transparency, 1 by default.}

\item{timeStartShapeVar, timeEndShapeVar}{(optional) string with
names of the variables in \code{data} used for symbol shape
for \code{timeStartVar}/\code{timeEndVar}.}

\item{shapePalette}{Named vector with shape for \code{timeStartShapeVar}\code{timeEndShapeVar}.}

\item{shapeLab}{String with label for \code{timeStartShapeVar}\code{timeEndShapeVar}}

\item{shapeSize}{Size for symbols (only used if \code{timeStartShapeVar}/\code{timeEndShapeVar} is specified).}

\item{title}{string, title for the plot}

\item{label}{string, label used in the plot.
This label is used to report the name of the panel 
in the text when the plots are combined (e.g. if the plot is empty).}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{formatReport}{list with parameters used to specify the format of the report,
e.g. output of the \code{\link{subjectProfileReportFormat}} function}

\item{paging}{Logical, if TRUE (by default), the automatic 
paging is enabled, otherwise only one page is used.}
}
\value{
list of (across subjects) of list (across modules) of \code{\link[ggplot2]{ggplot2} objects}, 
also of class \code{subjectProfileIntervalPlot}, with additional metaData attributes containing
'\code{label}' and 'timeLim' and 'timeTrans' (if specified).
}
\description{
If no start and/or end date is available and \code{timeStartShapeVar},
\code{timeEndShapeVar} are not specified: specific arrows are created.
}
\details{
In case of missing values for the time start or end variables, they are replaced by:
\itemize{
\item{in case \code{timeLimData}, \code{timeLimStartVar} and \code{timeLimEndVar} are specified: }{
minimum/maximum values in the variable: \code{timeLimVar} in the data: \code{timeLimData}
for the specific subject (if available). If there are missing for a specific subject,
they are taken across subjects.}
\item{otherwise, depending on the imputation type (\code{timeImpType}): }{
\itemize{
\item{'minimal' (by default): }{
\itemize{
\item{if start end and time are missing: }{no imputation is done}
\item{if start time is missing and end time is not missing: }{
start time is imputed with end time, and status is set to 'Missing start'}
\item{if end time is missing and start time is not missing: }{
end time is imputed with start time, and status is set to 'Missing end'}
}}
\item{'data-based' (default in version < 1.0.0): }{
minimum/maximum values in the start/end time variables in the data are considered
for the specific subject (if available). If there are missing for a specific subject,
they are taken across subjects. If all time are missings, the range is set to 0 and Inf}
\item{'none': }{no imputation is done}
}}}
The time limits are the same across subjects, and set to:
\itemize{
\item{\code{timeLim} if specified}
\item{maximum time range in \code{timeLimStartVar} and \code{timeLimEndVar} in \code{timeLimData} 
if specified}
\item{the maximum range on the data obtained after imputation of missing values}
}
}
\author{
Laure Cougnaud
}
