% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileLinePlot.R
\name{subjectProfileLinePlot}
\alias{subjectProfileLinePlot}
\title{Create spaghetti plot for subject profiles}
\usage{
subjectProfileLinePlot(data, paramValueVar,
  paramLab = toString(getLabelVar(paramValueVar, labelVars = labelVars)),
  paramNameVar = NULL, colorVar = NULL, colorLab = getLabelVar(colorVar,
  labelVars = labelVars), paramGroupVar = NULL, timeVar,
  subjectVar = "USUBJID", xLab = getLabelVar(timeVar, labelVars =
  labelVars), yLab = "", timeLim = NULL, title = paramLab,
  label = title, labelVars = NULL)
}
\arguments{
\item{data}{data.frame with data}

\item{paramValueVar}{string, variable of \code{data} with parameter value to represent}

\item{paramLab}{string, label for \code{paramVar}}

\item{paramNameVar}{string, variable of \code{data} with parameter name}

\item{colorVar}{string, variable of \code{data} with color.}

\item{colorLab}{string, label for \code{colorVar}}

\item{paramGroupVar}{(optional) character vector with variable(s) of \code{data} with grouping.
If specified, the parameters will be grouped by this(these) variable(s) in the y-axis.}

\item{timeVar}{string, variable of \code{data} with time}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{xLab}{string, label for the x-axis}

\item{yLab}{string, label for the y-axis}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} and maximum \code{timeEndVar}}

\item{title}{string, title for the plot}

\item{label}{string, label used in the plot.
This label is used to report the name of the panel 
in the text when the plots are combined (e.g. if the plot is empty).}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
list of \code{\link[ggplot2]{ggplot2} objects}, 
also of class \code{subjectProfileTextPlot}
}
\description{
Create spaghetti plot for subject profiles
}
\author{
Laure Cougnaud
}
