% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileRange.R
\name{subjectProfileRangePlot}
\alias{subjectProfileRangePlot}
\title{Create plot of subject profiles with segments for range of parameters}
\usage{
subjectProfileRangePlot(data, paramVar, paramLab = getLabelVar(paramVar,
  labelVars = labelVars), subjectVar, startVar, endVar, rangeLim = with(data,
  c(min(get(startVar), na.rm = TRUE), max(get(startVar), na.rm = TRUE))),
  xLab = paste(getLabelVar(c(startVar, endVar), labelVars = labelVars),
  collapse = "/"), yLab = "", colorVar = NULL,
  colorLab = getLabelVar(colorVar, labelVars = labelVars), title = paramLab,
  labelVars = NULL)
}
\arguments{
\item{data}{data.frame with data}

\item{paramVar}{variable of \code{data} with parameter (used in the y-axis)}

\item{paramLab}{label for \code{paramVar}}

\item{subjectVar}{variable of \code{data} with subject ID}

\item{startVar}{variable of \code{data} with start of range}

\item{endVar}{variable of \code{data} with end of range}

\item{rangeLim}{vector of length 2 with limits for the range (x-axis).
If not specified, extracted from the minimum \code{startVar} and maximum \code{endVar}}

\item{xLab}{label for the x-axis}

\item{yLab}{label for the y-axis}

\item{colorVar}{variable of \code{data} with color}

\item{colorLab}{label for \code{colorVar}}

\item{title}{title for the plot}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
list of \code{\link[ggplot2]{ggplot2} objects}
}
\description{
Create plot of subject profiles with segments for range of parameters
}
\author{
Laure Cougnaud
}
