% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{subjectProfileReportFormat}
\alias{subjectProfileReportFormat}
\title{Get list with format specification for subject profile report,
setting default for entire workflow}
\usage{
subjectProfileReportFormat(heightLineIn = 0.2, margin = 0.75,
  landscape = FALSE, aspectRatio = 0.5)
}
\arguments{
\item{heightLineIn}{height of a line in inches}

\item{margin}{margin in inches}

\item{landscape}{logical, if TRUE the created report is in landscape format}

\item{aspectRatio}{ratio between size of image in inches 
(derived from specified margin, landscape and heightLineIn)
and real size for exported image}
}
\value{
list with input parameters.
If not specified, default are used.
}
\description{
Get list with format specification for subject profile report,
setting default for entire workflow
}
\author{
Laure Cougnaud
}
