% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileSpaghettiPlot.R
\name{subjectProfileSpaghettiPlot}
\alias{subjectProfileSpaghettiPlot}
\title{Create spaghetti plot for subject profiles}
\usage{
subjectProfileSpaghettiPlot(data, paramVar,
  paramLab = toString(getLabelVar(paramVar, labelVars = labelVars)),
  facetVar = NULL, timeVar, subjectVar = "USUBJID",
  xLab = getLabelVar(timeVar, labelVars = labelVars), yLab = "",
  timeLim = NULL, title = paramLab, label = title, labelVars = NULL)
}
\arguments{
\item{data}{data.frame with data}

\item{paramVar}{string, variable of \code{data} with parameter (used in the y-axis)}

\item{paramLab}{string, label for \code{paramVar}}

\item{facetVar}{string, variable used for facetting}

\item{timeVar}{string, variable of \code{data} with time}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{xLab}{string, label for the x-axis}

\item{yLab}{string, label for the y-axis}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} and maximum \code{timeEndVar}}

\item{title}{string, title for the plot}

\item{label}{string, label used in the plot.
This label is used to report the name of the panel 
in the text when the plots are combined (e.g. if the plot is empty).}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
list of \code{\link[ggplot2]{ggplot2} objects}, 
also of class \code{subjectProfileTextPlot}
}
\description{
Create spaghetti plot for subject profiles
}
\author{
Laure Cougnaud
}
