% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileTextPlot.R
\name{subjectProfileTextPlot}
\alias{subjectProfileTextPlot}
\title{Create plot of subject profiles for events}
\usage{
subjectProfileTextPlot(data, vars, paramVar, paramValueVar,
  subjectVar = "USUBJID", xLab = "", yLab = "",
  title = "Subject information", labelVars = NULL)
}
\arguments{
\item{data}{data.frame with data}

\item{vars}{vector with string names of variables of \code{data} to consider}

\item{paramVar}{string, variable of \code{data} with parameter (used in the y-axis)}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{xLab}{string, label for the x-axis}

\item{yLab}{string, label for the y-axis}

\item{title}{string, title for the plot}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
list of \code{\link[ggplot2]{ggplot2} objects}
}
\description{
Create plot of subject profiles for events
}
\author{
Laure Cougnaud
}
