% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_subjectProfileCombine.R
\name{checkTimeTrans}
\alias{checkTimeTrans}
\title{Check if the subject profiles are time transformed,
and if some of the plots to align (with specified \code{timeLim})
have compatible time transformation alignments.}
\usage{
checkTimeTrans(listPlots, timeLim = NULL)
}
\arguments{
\item{listPlots}{list of plots}

\item{timeLim}{time limits}
}
\value{
List of time transformation for each module
(named by \code{listPlots})
}
\description{
Check if the subject profiles are time transformed,
and if some of the plots to align (with specified \code{timeLim})
have compatible time transformation alignments.
}
\author{
Laure Cougnaud
}
