% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{convertAesVar}
\alias{convertAesVar}
\title{Convert aesthetic variable for patient profile visualization.}
\usage{
convertAesVar(data, var)
}
\arguments{
\item{data}{data.frame with data}

\item{var}{variable of \code{data} with aesthetic}
}
\value{
updated factor \code{var} variable
}
\description{
This converts the empty values ('') to NA.
The variable is then converted as a factor.\cr
Missing values are also included in the 
levels of the factor, to ensure that missing values
are displayed in the legend of the plot.
}
\author{
Laure Cougnaud
}
