% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_createSubjectProfileReport.R
\name{createSubjectProfileReport}
\alias{createSubjectProfileReport}
\title{Create subject profile report.}
\usage{
createSubjectProfileReport(listPlots, timeLim = NULL,
  timeAlign = "all", timeAlignPerSubject = "none", refLines = NULL,
  refLinesData = NULL, refLinesTimeVar = NULL,
  refLinesLabelVar = NULL, bookmarkData = NULL, bookmarkVar = NULL,
  subjectSortData = bookmarkData, subjectSortVar = bookmarkVar,
  subjectVar = "USUBJID", subjectSubset = NULL,
  subjectSubsetData = NULL, subjectSubsetVar = NULL,
  subjectSubsetValue = NULL, subset = NULL,
  outputFile = "subjectProfile.pdf", exportFigures = FALSE,
  reportPerSubject = FALSE, labelVars = NULL, maxNLines = NULL,
  shiny = FALSE, formatReport = subjectProfileReportFormat(),
  verbose = FALSE, nCores = 1)
}
\arguments{
\item{listPlots}{nested list of plots, as returned by the \code{\link{subjectProfileTextPlot}},
\code{\link{subjectProfileEventPlot}}, \code{\link{subjectProfileIntervalPlot}} or
\code{\link{subjectProfileLinePlot}} functions.}

\item{timeLim}{Time limits, as a numeric vector of length 2,
or a list with time limits for each module,
or nested list with time limits for each module and subject.
If not specified, these are set to the time limits specified
when creating each module (stored in \code{attributes(x)$metaData$timeLim})
otherwise to the range defined by \code{timeAlign} and \code{timeAlignPerSubject}.
Note that this doesn't modify the geoms of the plots, it only extends the
axis range. So for interval module(s) if the specified \code{timeLim}
is smaller than the time limits in the input plot, no arrows are created in case than
the time goes above/below specified \code{timeLim} (the segment is cut).}

\item{timeAlign}{Character vector with time alignment, either:
\itemize{
\item{'all' (by default): }{all plots have the same time limits}
\item{'none': }{each of the plot has its own time limits}
\item{character vector with names of the modules which
should have the same time limits
(should correspond to the names of \code{listPlots})}
}}

\item{timeAlignPerSubject}{Character vector specifying
which modules of \code{timeAlign} should have subject-specific time limits
\itemize{
\item{'none' (by default): }{these modules have the same time limit across subjects}
\item{'all': }{these modules have different time limits per subject}
\item{character vector with subset of these modules which should have
 different time limits per subject
(should correspond to the names of \code{listPlots})}
}
This is only used for the modules with which \code{timeAlign} is specified.}

\item{refLines}{(optional) nested list with details for reference line(s).
Each sublist contains:
\itemize{
\item{(required) 'label': }{string with label for the reference line}
\item{(required) 'time': }{time(x) coordinate for the reference line,
'dotted' by default}
\item{(optional) 'color': }{color for the reference line,
'black' by default}
\item{(optional) 'linetype': }{linetype for the reference line}
}}

\item{refLinesData}{data.frame with data from which the reference line(s) should be extracted}

\item{refLinesTimeVar}{string, variable of \code{refLinesData} with time for reference line(s)}

\item{refLinesLabelVar}{string, variable of \code{refLinesData} with label for reference line(s)}

\item{bookmarkData}{data.frame with data containing information on which the index should be based}

\item{bookmarkVar}{variable(s) of \code{data} of interest for the index}

\item{subjectSortData}{data.frame with data containing information on how the subjects 
should be sorted in the report, by default same as \code{bookmarkData}}

\item{subjectSortVar}{variable(s) of \code{data} indicating the order for the subjects in the report,
by default same as \code{bookmarkVar}}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{subjectSubset}{subjectSubset (optional) Character vector with subjects of interest 
(available in \code{subjectVar}), NULL by default.}

\item{subjectSubsetData}{data.frame with data used to select subset of subjects of interest.
It should contain the \code{subjectVar} variable.}

\item{subjectSubsetVar}{string with variable of \code{subjectSubsetData} used for subsetting.}

\item{subjectSubsetValue}{Character vector with value(s) of \code{subjectSubsetVar}
of interest to select subjects on.}

\item{subset}{Character vector with subjects of interest 
(among names of each list in \code{listPlots}).}

\item{outputFile}{string, path to the output report}

\item{exportFigures}{logical, if TRUE (FALSE by default) the figures are also exported
in png format in a 'figures' folder}

\item{reportPerSubject}{Logical, if TRUE (FALSE by default)
export a subject profile report by subject.}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{maxNLines}{maximum number of lines for a combined plot,
to fit in the page height}

\item{shiny}{logical, set to TRUE (FALSE by default) 
if the report is generated from a Shiny application.
Messages during report creation will be included in the Shiny interface,
and it will be mentioned at the end of the report.
In this case, the \code{shiny} package should be available.}

\item{formatReport}{list with parameters used to specify the format of the report,
e.g. output of the \code{\link{subjectProfileReportFormat}} function}

\item{verbose}{logical, if TRUE print messages during execution}

\item{nCores}{Integer containing the number of cores used for the computation
(1 by default). If more than 1, computation is parallelized, in this case
the package \code{parallel} is required.}
}
\value{
The path(s) of the report(s) is returned invisibly, and the
the report is created at the location
specified by \code{outputFile}.
}
\description{
By default all subjects available in at least one module of \code{listPlots} are considered.
If only a set of subjects are of interest, these are specified either:
\itemize{
\item{directly with the subject IDs of interest via \code{subjectSubset}}
\item{by extracting subjects with a specific value (\code{subjectSubsetValue})
in a variable (\code{subjectSubsetVar}) in a specific dataset \code{subjectSubsetData}}
}
}
\author{
Laure Cougnaud
}
