% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfileReport.R
\name{createSubjectProfileReport}
\alias{createSubjectProfileReport}
\title{Create subject profile report}
\usage{
createSubjectProfileReport(listPlots,
  timeLim = getXLimSubjectProfilePlots(listPlots), refLines = NULL,
  refLinesData = NULL, refLinesTimeVar = NULL, refLinesLabelVar = NULL,
  subjectVar = "USUBJID", landscape = FALSE,
  outputFile = "subjectProfile.pdf", exportFigures = FALSE)
}
\arguments{
\item{listPlots}{list of plots, as returned by the \code{\link{subjectProfileTextPlot}},
\code{\link{subjectProfileEventPlot}} or \code{\link{subjectProfileIntervalPlot}}}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{startVar} and maximum \code{endVar}}

\item{refLines}{(optional) nested list with details for reference line(s).
Each sublist contains:
\itemize{
\item{(required) 'label': }{string with label for the reference line}
\item{(required) 'time': }{time(x) coordinate for the reference line,
'dotted' by default}
\item{(optional) 'color': }{color for the reference line,
'black' by default}
\item{(optional) 'linetype': }{linetype for the reference line}
}}

\item{refLinesData}{data.frame with data from which the reference line(s) should be extracted}

\item{refLinesTimeVar}{string, variable of \code{refLinesData} with time for reference line(s)}

\item{refLinesLabelVar}{string, variable of \code{refLinesData} with label for reference line(s)}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{landscape}{logical, if TRUE the created report is in landscape format}

\item{outputFile}{string, path to the output report}

\item{exportFigures}{logical, if TRUE (FALSE by default) the figures are also exported
in png format in a 'figures' folder}
}
\value{
no returned value, the report is created at the location
specified by \code{outputFile}
}
\description{
Create subject profile report
}
\author{
Laure Cougnaud
}
