% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfileReport.R
\name{createSubjectProfileReport}
\alias{createSubjectProfileReport}
\title{Create subject profile report}
\usage{
createSubjectProfileReport(listPlots,
  timeLim = getXLimSubjectProfilePlots(listPlots), refLines = NULL,
  refLinesData = NULL, refLinesTimeVar = NULL, refLinesLabelVar = NULL,
  bookmarkData = NULL, bookmarkVar = NULL, subjectVar = "USUBJID",
  landscape = FALSE, outputFile = "subjectProfile.pdf",
  exportFigures = FALSE, labelVars = NULL, heightLineIn = 0.2)
}
\arguments{
\item{listPlots}{list of plots, as returned by the \code{\link{subjectProfileTextPlot}},
\code{\link{subjectProfileEventPlot}} or \code{\link{subjectProfileIntervalPlot}}}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} and maximum \code{timeEndVar}}

\item{refLines}{(optional) nested list with details for reference line(s).
Each sublist contains:
\itemize{
\item{(required) 'label': }{string with label for the reference line}
\item{(required) 'time': }{time(x) coordinate for the reference line,
'dotted' by default}
\item{(optional) 'color': }{color for the reference line,
'black' by default}
\item{(optional) 'linetype': }{linetype for the reference line}
}}

\item{refLinesData}{data.frame with data from which the reference line(s) should be extracted}

\item{refLinesTimeVar}{string, variable of \code{refLinesData} with time for reference line(s)}

\item{refLinesLabelVar}{string, variable of \code{refLinesData} with label for reference line(s)}

\item{bookmarkData}{data.frame with data containing information on which the index should be based}

\item{bookmarkVar}{variable(s) of \code{data} of interest for the index}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{landscape}{logical, if TRUE the created report is in landscape format}

\item{outputFile}{string, path to the output report}

\item{exportFigures}{logical, if TRUE (FALSE by default) the figures are also exported
in png format in a 'figures' folder}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{heightLineIn}{height of a line in inches}
}
\value{
no returned value, the report is created at the location
specified by \code{outputFile}
}
\description{
Create subject profile report
}
\author{
Laure Cougnaud
}
