% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfileReport.R
\name{createSubjectProfileReport}
\alias{createSubjectProfileReport}
\title{Create subject profile report}
\usage{
createSubjectProfileReport(listPlots,
  timeLim = getXLimSubjectProfilePlots(listPlots), refLines = NULL,
  refLinesData = NULL, refLinesTimeVar = NULL, refLinesLabelVar = NULL,
  bookmarkData = NULL, bookmarkVar = NULL, subjectSortData = bookmarkData,
  subjectSortVar = bookmarkVar, subjectVar = "USUBJID",
  outputFile = "subjectProfile.pdf", exportFigures = FALSE,
  labelVars = NULL, maxNLines = NULL, shiny = FALSE,
  formatReport = subjectProfileReportFormat())
}
\arguments{
\item{listPlots}{nested list of plots, as returned by the \code{\link{subjectProfileTextPlot}},
\code{\link{subjectProfileEventPlot}}, \code{\link{subjectProfileIntervalPlot}} or
\code{\link{subjectProfileLinePlot}} functions.}

\item{timeLim}{vector of length 2 with time limits.
If not specified, these are set to the time limits specified
when creating each module (stored in \code{attributes(x)$metaData$timeLim})
otherwise to the maximum range contained in the data.}

\item{refLines}{(optional) nested list with details for reference line(s).
Each sublist contains:
\itemize{
\item{(required) 'label': }{string with label for the reference line}
\item{(required) 'time': }{time(x) coordinate for the reference line,
'dotted' by default}
\item{(optional) 'color': }{color for the reference line,
'black' by default}
\item{(optional) 'linetype': }{linetype for the reference line}
}}

\item{refLinesData}{data.frame with data from which the reference line(s) should be extracted}

\item{refLinesTimeVar}{string, variable of \code{refLinesData} with time for reference line(s)}

\item{refLinesLabelVar}{string, variable of \code{refLinesData} with label for reference line(s)}

\item{bookmarkData}{data.frame with data containing information on which the index should be based}

\item{bookmarkVar}{variable(s) of \code{data} of interest for the index}

\item{subjectSortData}{data.frame with data containing information on how the subjects 
should be sorted in the report, by default same as \code{bookmarkData}}

\item{subjectSortVar}{variable(s) of \code{data} indicating the order for the subjects in the report,
by default same as \code{bookmarkVar}}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{outputFile}{string, path to the output report}

\item{exportFigures}{logical, if TRUE (FALSE by default) the figures are also exported
in png format in a 'figures' folder}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{maxNLines}{maximum number of lines for a combined plot,
to fit in the page height}

\item{shiny}{logical, set to TRUE (FALSE by default) if the report is generated from a Shiny application.
Messages during report creation will be included in the Shiny interface,
and it will be mentioned at the end of the report.
In this case, the \code{shiny} package should be available.}

\item{formatReport}{list with parameters used to specify the format of the report,
see parameters of the \code{\link{subjectProfileReportFormat}} function}
}
\value{
no returned value, the report is created at the location
specified by \code{outputFile}
}
\description{
Create subject profile report
}
\author{
Laure Cougnaud
}
