% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyEventPlot.R
\name{eventPlotHoverTemplate}
\alias{eventPlotHoverTemplate}
\title{Create [plotly] tool tip hover template for a event plot}
\usage{
eventPlotHoverTemplate(
  data,
  paramVar,
  paramLab,
  timeVar,
  timeLab,
  colorLab = NULL,
  colorVar = NULL,
  shapeLab = NULL,
  shapeVar = NULL,
  add_vars = NULL
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{paramVar}{Character vector with variable(s) of \code{data} 
with parameters. Variable content is displayed in the y-axis.}

\item{paramLab}{Named character vector, 
with label for the parameter variable(s) (\code{paramVar}).\cr
This is used to set the default title.}

\item{timeVar}{String, variable of \code{data} with time,
displayed in the x axis.\cr
Records with missing time are not displayed in the plot.}

\item{timeLab}{String, label for \code{timeVar}.
This is used in the message
indicating missing values for \code{timeVar},
and for the default label of the x-axis.}

\item{colorLab}{String, label for \code{colorVar}.}

\item{colorVar}{String, variable of \code{data} with color.}

\item{shapeLab}{String, label for \code{shapeVar}.
Set by default to \code{colorLab} if \code{colorVar}
but not \code{shapeVar} is not specified.}

\item{shapeVar}{String, variable of \code{data} for shape of the points.
By default, same as \code{colorVar}.}

\item{add_vars}{see [subjectProfileLinePlot()] argument `plotly_add_vars`}
}
\value{
a character vector of html formatted strings
}
\description{
Create [plotly] tool tip hover template for a event plot
}
