% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{filterData}
\alias{filterData}
\title{Filter a dataset for records of interest}
\usage{
filterData(data, subsetVar = NULL, subsetValue = NULL)
}
\arguments{
\item{data}{data data.frame with data}

\item{subsetVar}{variable of \code{data} used for subsetting}

\item{subsetValue}{character vector with value(s) of interest to consider for 
\code{subsetVar}}
}
\value{
possibly filtered dataset
}
\description{
Filter a dataset for records of interest
}
\author{
Laure Cougnaud
}
