% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{formatParamVar}
\alias{formatParamVar}
\title{Format a variable as factor, and optionally sorted according to the levels
of a grouping variable}
\usage{
formatParamVar(data, paramVar = NULL, paramGroupVar = NULL,
  revert = FALSE, width = 20)
}
\arguments{
\item{data}{data.frame with data}

\item{paramVar}{string, variable of \code{data} with parameter}

\item{paramGroupVar}{(optional) character vector with variable(s) of \code{data} with grouping.
If specified, the parameters will be grouped by this(these) variable(s) in the y-axis.}

\item{revert}{logical, if TRUE revert the order of the levels of the variable}

\item{width}{max}
}
\value{
vector with re-formatted \code{paramVar}, NULL if empty
}
\description{
Format a variable as factor, and optionally sorted according to the levels
of a grouping variable
}
\author{
Laure Cougnaud
}
