% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileIntervalPlot.R
\name{formatTimeInterval}
\alias{formatTimeInterval}
\title{Set missing start/end time variable in the data.}
\usage{
formatTimeInterval(
  data,
  timeStartVar,
  timeStartLab = getLabelVar(timeStartVar, labelVars = labelVars),
  timeEndVar,
  timeEndLab = getLabelVar(timeEndVar, labelVars = labelVars),
  subjectVar = "USUBJID",
  timeLim = NULL,
  timeLimData = NULL,
  timeLimStartVar = NULL,
  timeLimStartLab = getLabelVar(timeLimStartVar, labelVars = labelVars),
  timeLimEndVar = NULL,
  timeLimEndLab = getLabelVar(timeLimEndVar, labelVars = labelVars),
  timeImpType = c("minimal", "data-based", "none"),
  labelVars = NULL
)
}
\arguments{
\item{data}{data.frame with data}

\item{timeStartVar}{string, variable of \code{data} with start of range}

\item{timeStartLab}{String, label for \code{timeStartVar}.}

\item{timeEndVar}{string, variable of \code{data} with end of range}

\item{timeEndLab}{String, label for \code{timeEndVar}.}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} 
and maximum \code{timeEndVar} per subject.}

\item{timeLimData}{data.frame with data used to extract time limits per subject}

\item{timeLimStartVar}{string, variable of \code{timeLimData} with time start}

\item{timeLimEndVar}{string, variable of \code{timeLimData} with time end}

\item{timeImpType}{String with imputation type: 'minimal' (default),
'data-based' or 'none'.}
}
\value{
list with:
\itemize{
\item{'data': }{Data with:
\itemize{
\item{imputed \code{timeStartVar} and \code{timeEndVar}}
\item{new column 'timeStartStatus': }{
character vector containing status of \code{timeStartVar} variable:
 'Complete' or 'Missing start' or NA}
\item{new column 'timeEndStatus': }{
character vector containing status of \code{timeEndVar} variable:
 'Complete' or 'Missing end' or NA}
}}
\item{'timeLim': }{vector of length 2 with minimum/maximum time limits across subjects.}
\item{'timeLimSpecified': }{vector of length 2 with time limits as specified by the user,
either extracted from \code{timeLim} or from \code{timeLimData}.
If missing value within \code{timeLim}, the corresponding minimum/maximum
value in the (updated) data is used.}
\item{'timeShapePalette': }{Named character vector with symbols for the different time status}
\item{'caption': }{String with extra explanation concerning imputation that could be included in plot caption.}
}
}
\description{
Set missing start/end time variable in the data.
}
\details{
In case of missing values for the time start or end variables, they are replaced by:
\itemize{
\item{in case \code{timeLimData}, \code{timeLimStartVar} and \code{timeLimEndVar} are specified: }{
 the minimum/maximum values in the variables: \br\code{timeLimStartVar}/\code{timeLimEndVar} 
respectively are considered
 in the data: \code{timeLimData}
for the specific subject (if available). If there are missing for a specific subject,
they are taken across subjects.}
\item{otherwise, depending on the imputation type (\code{timeImpType}): }{
\itemize{
\item{'minimal' (by default): }{
\itemize{
\item{if start end and time are missing: }{no imputation is done, only the label is displayed}
\item{if start time is missing and end time is not missing: }{
start time is imputed with end time, and status is set to 'Missing start'}
\item{if end time is missing and start time is not missing: }{
end time is imputed with start time, and status is set to 'Missing end'}
}}
\item{'data-based' (default in version < 1.0.0): }{
minimum/maximum values in the start/end time variables in the data are considered
for the specific subject (if available). If there are missing for a specific subject,
they are taken across subjects. If all time are missings, the range is set to 0 and Inf}
\item{'none': }{no imputation is done}
}}}
The time limits are the same across subjects, and set to:
\itemize{
\item{\code{timeLim} if specified}
\item{maximum time range in \code{timeLimStartVar} and \code{timeLimEndVar} in \code{timeLimData} 
if specified}
\item{the maximum range on the data obtained after imputation of missing values}
}
}
\author{
Laure Cougnaud
}
