% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileIntervalPlot.R
\name{formatTimeInterval}
\alias{formatTimeInterval}
\title{Set missing start/end time variable in the data.}
\usage{
formatTimeInterval(data, timeStartVar, timeEndVar, subjectVar = "USUBJID",
  timeLim = NULL, timeLimData = NULL, timeLimStartVar = NULL,
  timeLimEndVar = NULL)
}
\arguments{
\item{data}{data.frame with data}

\item{timeStartVar}{string, variable of \code{data} with start of range}

\item{timeEndVar}{string, variable of \code{data} with end of range}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} 
and maximum \code{timeEndVar} per subject.}

\item{timeLimData}{data.frame with data used to extract time limits per subject}

\item{timeLimStartVar}{string, variable of \code{timeLimData} with time start}

\item{timeLimEndVar}{string, variable of \code{timeLimData} with time end}

\item{timeLimInData}{vector of length 2 with time limits in the data (across subjects).}
}
\value{
list with:
\itemize{
\item{'data': }{data with filled missing start/end time variables.
The data also contains additional logical columns 'missingStartPlot' and 'missingEndPlot',
set to TRUE if the records had missing \code{timeStartVar} and \code{timeEndVar} variables.
}
\item{'timeLim': }{vector of length 2 with minimum/maximum time limits across subjects.
}
}
}
\description{
Set missing start/end time variable in the data.
}
\details{
In case of missing values for the time start or end variables, they are replaced by:
\itemize{
\item{in case \code{timeLim} is specified: }{specified time limits}
\item{otherwise, in case \code{timeLimData} and \code{timeLimVar} are specified: }{
minimum/maximum values in the variable: \code{timeLimVar} in the data: \code{timeLimData}
for the specific subject (if available). If there are missing for a specific subject,
they are taken across subjects.
}
\item{otherwise (default): }{the missing values for start/end date are replaced by the 
minimum/maximum of the times available in the \code{timeStartVar} and 
\code{timeEndVar} of the visualized \code{data} PER SUBJECT (if available).
}
\item{if all time are missings, the range is set to 0 and Inf} 
}
The time limits are the same across subjects, and set to:
1) \code{timeLim} if specified
2) maximum time range in \code{timeLimStartVar} and \code{timeLimEndVar} in \code{timeLimData} 
if specified
3) the maximum range on the data obtained after filling missing values
}
\author{
Laure Cougnaud
}
