% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{getGLPGShapePalettePatientProfile}
\alias{getGLPGShapePalettePatientProfile}
\title{Get a Galapagos shape palette palette of specified length
for patient profile visualization visualization.
This is only a wrapper above \code{\link[glpgUtilityFct]{getGLPGShapePalette}},
including by default missing values (contrary to the default of the
code{\link[glpgUtilityFct]{getGLPGShapePalette}} function).}
\usage{
getGLPGShapePalettePatientProfile(..., includeNA = TRUE)
}
\arguments{
\item{...}{Additional parameters for the 
\code{\link[glpgUtilityFct]{getGLPGShapePalette}} function.}

\item{includeNA}{Logical (TRUE by default), should NA elements be retained in the palette.}
}
\value{
vector of integer values with shape
}
\description{
Get a Galapagos shape palette palette of specified length
for patient profile visualization visualization.
This is only a wrapper above \code{\link[glpgUtilityFct]{getGLPGShapePalette}},
including by default missing values (contrary to the default of the
code{\link[glpgUtilityFct]{getGLPGShapePalette}} function).
}
\author{
Laure Cougnaud
}
