% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getMaxNLinesCombinePlot}
\alias{getMaxNLinesCombinePlot}
\title{Get maximum number of lines of a 'combined plot'
for a specific document}
\usage{
getMaxNLinesCombinePlot(
  heightLineIn = subjectProfileReportFormat()$heightLineIn,
  margin = subjectProfileReportFormat()$margin,
  landscape = subjectProfileReportFormat()$landscape,
  aspectRatio = subjectProfileReportFormat()$aspectRatio
)
}
\arguments{
\item{heightLineIn}{height of a line in inches}

\item{margin}{margin in inches}

\item{landscape}{logical, if TRUE the created report is in landscape format}

\item{aspectRatio}{ratio between size of image in inches 
(derived from specified margin, landscape and heightLineIn)
and real size for exported image}
}
\value{
numeric with maximum height for plot
}
\description{
Get maximum number of lines of a 'combined plot'
for a specific document
}
\author{
Laure Cougnaud
}
