% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getMaxNLinesCombinePlot}
\alias{getMaxNLinesCombinePlot}
\title{Get maximum number of lines of a 'combined plot'
for a specific document}
\usage{
getMaxNLinesCombinePlot(
  heightLineIn = subjectProfileReportFormat()$heightLineIn,
  margin = subjectProfileReportFormat()$margin,
  landscape = subjectProfileReportFormat()$landscape,
  aspectRatio = subjectProfileReportFormat()$aspectRatio
)
}
\arguments{
\item{heightLineIn}{Numeric of length 1 with 
height of a line in inches, 0.2 by default.}

\item{margin}{Numeric of length 1, with margin in inches.}

\item{landscape}{Logical, if TRUE the created report is 
in landscape format. FALSE by default, the report
is created in portrait format.}

\item{aspectRatio}{Numeric of length 1 (0.75 by default)
with ratio between size of image in inches 
(derived from specified \code{margin}, \code{landscape} and \code{heightLineIn})
and real size for exported image.}
}
\value{
numeric with maximum height for plot
}
\description{
Get maximum number of lines of a 'combined plot'
for a specific document
}
\author{
Laure Cougnaud
}
