% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getPageVar}
\alias{getPageVar}
\title{Get variable with page of the plot,
used for automatic paging of a plot}
\usage{
getPageVar(data, var, typeVar = c("y", "panel"),
  formatReport = subjectProfileReportFormat(), title = TRUE, xLab = TRUE,
  caption = TRUE)
}
\arguments{
\item{data}{data.frame with data}

\item{var}{string, variable of \code{data} with variable for the y-axis}

\item{typeVar}{string, type of the variable, either 'y' or 'panel'}

\item{formatReport}{list of output from the 
\code{\link{subjectProfileReportFormat}} function}

\item{title}{logical, has the plot a title?}

\item{xLab}{logical, has the plot a label for the x-axis?}

\item{caption}{logical, has the plot a caption?}
}
\value{
input \code{data} with additional column 'pagePlot'
containing the page for the plot
}
\description{
Get variable with page of the plot,
used for automatic paging of a plot
}
\author{
Laure Cougnaud
}
