% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getWidthPlot}
\alias{getWidthPlot}
\title{Get width for a plot for a certain page layout}
\usage{
getWidthPlot(
  margin = subjectProfileReportFormat()$margin,
  landscape = subjectProfileReportFormat()$landscape,
  aspectRatio = subjectProfileReportFormat()$aspectRatio
)
}
\arguments{
\item{margin}{Numeric of length 1, with margin in inches.}

\item{landscape}{Logical, if TRUE the created report is 
in landscape format. FALSE by default, the report
is created in portrait format.}

\item{aspectRatio}{Numeric of length 1 (0.75 by default)
with ratio between size of image in inches 
(derived from specified \code{margin}, \code{landscape} and \code{heightLineIn})
and real size for exported image.}
}
\value{
width for the plot in inches
}
\description{
Get width for a plot for a certain page layout
}
\author{
Laure Cougnaud
}
