% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_createSubjectProfileReport.R
\name{getXLimSubjectProfilePlots}
\alias{getXLimSubjectProfilePlots}
\title{Get limits for a list of plots.}
\usage{
getXLimSubjectProfilePlots(listPlots, align = TRUE)
}
\arguments{
\item{listPlots}{list of list of \code{subjectProfile[X]Plot} plots}

\item{align}{Logical, if TRUE (by default) the plots are aligned,
the limits of all plots are combined to extract the limits.}
}
\value{
vector of length 2 with limits for the x-axis
}
\description{
These limits are extracted from specified \code{timeLim} for each
module (stored in the \code{attributes()$metaData$timeLim}), 
and if empty for all modules: from the maximal range
of the x-coordinates across all plots.
}
\author{
Laure Cougnaud
}
