% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyIntervalPlot.R
\name{intervalPlotHoverTemplate}
\alias{intervalPlotHoverTemplate}
\title{Create [plotly] tool tip hover template for an interval plot}
\usage{
intervalPlotHoverTemplate(
  paramVal,
  paramLab,
  timeStartVal,
  timeStartLab,
  timeEndVal,
  timeEndLab,
  colorLab = NULL,
  colorVal = NULL,
  add_vars = NULL
)
}
\arguments{
\item{paramVal}{the value of `paramVar` for the tooltip}

\item{paramLab}{Named character vector, 
with label for the parameter variable(s) (\code{paramVar}).\cr
This is used to set the default title.}

\item{timeStartVal}{the value of `timeStartVar` for the tooltip}

\item{timeStartLab}{String, label for \code{timeStartVar},
displayed in a message and in the plot caption.}

\item{timeEndVal}{the value of `timeEndVar` for the tooltip}

\item{timeEndLab}{String, label for \code{timeEndVar},
displayed in a message and in the plot caption.}

\item{colorLab}{String, label for \code{colorVar}.}

\item{colorVal}{the value of `colorVar` for the tooltip}

\item{add_vars}{see [subjectProfileLinePlot()] argument `plotly_args`}
}
\value{
a string with html formatting
}
\description{
Create [plotly] tool tip hover template for an interval plot
}
