% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labelVarsPelican}
\alias{labelVarsPelican}
\title{Variable labels extracted from the \code{sdtm} datasets for the
Pelican study ('study').}
\format{named vector with label for each variable (names)}
\usage{
labelVarsPelican
}
\description{
This includes the  adverse event ('ae'), medical history ('mh'), 
demographics ('dm'), and laboratory ('lb') data.
}
\keyword{datasets}
