% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labelVarsPelican}
\alias{labelVarsPelican}
\title{Variable labels extracted from the \code{sdtm} datasets for the
Pelican study ('study').
This corresponds to the \code{\link{sdtmDataPelican}} dataset.}
\format{named vector with label for each variable (names)}
\description{
Variable labels extracted from the \code{sdtm} datasets for the
Pelican study ('study').
This corresponds to the \code{\link{sdtmDataPelican}} dataset.
}
\author{
Laure Cougnaud
}
\keyword{data}
