% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labelVarsSDTMPelican}
\alias{labelVarsSDTMPelican}
\title{Variable labels extracted from the \code{SDTM} datasets for the
Pelican study ('GLPG2737-CL-202-UNBLINDED/CRG10GLP01_SDTM_2018_06JUN_15_v2'), 
extracted from the \code{\link{SDTMDataPelican}} dataset.}
\format{named vector with label for each variable (names)}
\description{
Variable labels extracted from the \code{SDTM} datasets for the
Pelican study ('GLPG2737-CL-202-UNBLINDED/CRG10GLP01_SDTM_2018_06JUN_15_v2'), 
extracted from the \code{\link{SDTMDataPelican}} dataset.
}
\author{
Laure Cougnaud
}
\keyword{data}
