% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientProfilesVis-common-args-doc.R
\name{patientProfilesVis-common-args}
\alias{patientProfilesVis-common-args}
\title{Arguments used across the functions of the patientProfilesVis package.}
\arguments{
\item{data}{Data.frame with data.}

\item{xLab}{String, label for the x-axis.}

\item{yLab}{String, label for the y-axis.}

\item{label}{String, label for the visualization.
This label is used when the visualization is 
included in the final patient profile report,
in case this visualization is empty (no data available)
for a specific subject.}
}
\description{
Arguments used across the functions of the patientProfilesVis package.
}
