% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{patientProfilesVis-palette}
\alias{patientProfilesVis-palette}
\title{Parameters for all patient profiles visualization palette functions.}
\arguments{
\item{x}{Vector with elements used for palette.
If factor, the levels are used, otherwise the unique elements of the vector.
Missing values are automatically removed, excepted if
\code{includeNA} is set to TRUE.}

\item{includeNA}{Logical (TRUE by default), 
should NA elements be retained in the palette in case
\code{x} is specified?}

\item{n}{Integer of length 1, number of elements in palette.}
}
\description{
Parameters for all patient profiles visualization palette functions.
}
