% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotlyLinePlot}
\alias{plotlyLinePlot}
\title{Create a plotly line plot facetted by paramFacetVar}
\usage{
plotlyLinePlot(
  data,
  paramValueVar,
  paramValueVarUnits = NULL,
  paramValueRangeVar = NULL,
  colorValueRange,
  colorVar = NULL,
  colorLab = NULL,
  colorPalette,
  alpha = 1,
  shapeVar = NULL,
  shapeLab = NULL,
  shapePalette = NULL,
  shapeSize = 7,
  timeVar,
  timeLab,
  title,
  xLab,
  add_vars = NULL,
  facetVarMaxLength = 30,
  margin = list(l = 250, r = 250, b = 75, t = 50, pad = 4),
  yaxis_title_shift = -0.035,
  legend_x_shift = 1.2,
  spikecolor = "red"
)
}
\arguments{
\item{data}{a data frame with data for 1 subject only}

\item{paramValueVar}{String, variable of \code{data} with parameter value to
represent.\cr Records with missing values are discarded.}

\item{paramValueRangeVar}{Character vector of length 2 containing variables
of \code{data} with minimum and maximum value for \code{paramValueVar},
typically reference range indicators.\cr Range can differ per parameter and
even per time point. This range is represented as a ribbon in the plot
background. e.g. to represent the reference range of the variable.}

\item{colorValueRange}{String with color for the filling of the ribbon
represented by \code{paramValueRangeVar}.}

\item{colorVar}{String, variable of \code{data} with color.}

\item{colorLab}{String, label for \code{colorVar}.}

\item{colorPalette}{Named vector with color palette.
The variable should be named with the corresponding element
in \code{colorVar}.\cr
Colors can also be defined for the entire session, by setting
\code{options(patientProfilesVis.colors = X)} with X either:
\itemize{
\item{a vector with colors}
\item{a function returning a vector of colors for a specified number
of elements (\code{viridis} by default)}
}}

\item{alpha}{Numeric with transparency, 1 by default.}

\item{shapeVar}{String, variable of \code{data} for shape of the points.
By default, same as \code{colorVar}.}

\item{shapeLab}{String, label for \code{shapeVar}.
Set by default to \code{colorLab} if \code{colorVar}
but not \code{shapeVar} is not specified.}

\item{shapePalette}{Named character vector with shape palette
for \code{shapeVar}.
The variable should be named with the corresponding element
in \code{shapeVar}.\cr
Shapes can also be defined for the entire session, by setting
\code{options(patientProfilesVis.shapes = X)} with X either:
\itemize{
\item{a vector with shapes}
\item{a function returning a vector of shapes for a specified number
of elements}
}
Note it is advised to specify the shapes as character,
e.g. 'cross' instead of \code{4}, in case Unicode
symbols should also be used.}

\item{shapeSize}{Size for the symbols, any integer or object supported by
\code{size} in \code{\link[ggplot2]{geom_point}}. If `plotly = TRUE`, 
recommend setting this value to `7` or higher.}

\item{timeVar}{String, variable of \code{data} with time,
displayed in the x axis.\cr
Records with missing time are not displayed in the plot.}

\item{timeLab}{String, label for \code{timeVar}.
This is used in the message
indicating missing values for \code{timeVar},
and for the default label of the x-axis.}

\item{title}{String, title for the plot.}

\item{xLab}{String, label for the x-axis.}

\item{add_vars}{see [subjectProfileLinePlot()] argument `plotly_args`}

\item{facetVarMaxLength}{see [subjectProfileLinePlot()] argument
`plotly_args`}

\item{margin}{see [subjectProfileLinePlot()] argument `plotly_args`}

\item{yaxis_title_shift}{see [subjectProfileLinePlot()] argument
`plotly_args`}

\item{spikecolor}{see [subjectProfileLinePlot()] argument `plotly_args`}

\item{legend_y_shift}{see [subjectProfileLinePlot()] argument `plotly_args`}
}
\value{
a [plotly] object
}
\description{
Create a plotly line plot facetted by paramFacetVar
}
\details{
This function is designed to be called by [subjectProfileLinePlot()].
}
