% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SDTMDataPelican}
\alias{SDTMDataPelican}
\title{Example of SDTM data for subject profiles, currently 
extracted from the \code{SDTM} datasets for the
Pelican study ('study').}
\format{list with data.frame for each dataset}
\description{
This includes the  adverse event ('ae'), medical history ('mh'), 
exposure ('ex'), demographics ('dm'), and laboratory ('lb') data.
}
\author{
Laure Cougnaud
}
\keyword{data}
