% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfileReport.R
\name{subjectProfileCombine}
\alias{subjectProfileCombine}
\title{Combine subject profile plots.}
\usage{
subjectProfileCombine(listPlots)
}
\arguments{
\item{listPlots}{list of \code{\link[ggplot2]{ggplot2} objects}}
}
\value{
a list of \code{subjectProfilePlot} object, containing the combined
profile plots for each subject.
This is in essence only a \code{\link[gtable]{gtable}} object,
wrapped within a \code{subjectProfilePlot} class for dedicated
'print' function.
}
\description{
Combine subject profile plots.
}
\author{
Laure Cougnaud
}
