% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfileReport.R
\name{subjectProfileCombine}
\alias{subjectProfileCombine}
\title{Combine subject profile plots.}
\usage{
subjectProfileCombine(listPlots,
  timeLim = getXLimSubjectProfilePlots(listPlots))
}
\arguments{
\item{listPlots}{list of \code{\link[ggplot2]{ggplot2} objects}}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{startVar} and maximum \code{endVar}}
}
\value{
a list of \code{subjectProfilePlot} object, containing the combined
profile plots for each subject.
This is in essence only a \code{\link[ggplot2]{ggplot2}} objects,
(with the additional attribute 'nLinesYAxis')
}
\description{
Combine subject profile plots.
}
\author{
Laure Cougnaud
}
