% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_subjectProfileCombine.R
\name{subjectProfileCombine}
\alias{subjectProfileCombine}
\title{Combine subject profile plots.}
\usage{
subjectProfileCombine(listPlots, timeLim = NULL, timeAlign = "all",
  timeAlignPerSubject = "none", subjectVar = "USUBJID",
  maxNLines = NULL, refLines = NULL, refLinesData = NULL,
  refLinesTimeVar = NULL, refLinesLabelVar = NULL, shiny = FALSE,
  verbose = FALSE, nCores = 1)
}
\arguments{
\item{listPlots}{list of subject profiles (modules/subjects)
Messages during report creation will be included in the Shiny interface,
and it will be mentioned at the end of the report.
In this case, the \code{shiny} package should be available.}

\item{timeLim}{time limits, as returned by the
\code{\link{getXLimSubjectProfilePlots}} function.
If NULL, the original time limits of the plot are used.}

\item{timeAlign}{Logical, if TRUE (by default)
the different plots are horizontally aligned.
If set to FALSE, each plot has its own time-limits.}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{maxNLines}{maximum number of lines for a combined plot,
to fit in the page height}

\item{refLines}{(optional) nested list with details for reference line(s).
Each sublist contains:
\itemize{
\item{(required) 'label': }{string with label for the reference line}
\item{(required) 'time': }{time(x) coordinate for the reference line,
'dotted' by default}
\item{(optional) 'color': }{color for the reference line,
'black' by default}
\item{(optional) 'linetype': }{linetype for the reference line}
}}

\item{refLinesData}{data.frame with data from which the reference line(s) should be extracted}

\item{refLinesTimeVar}{string, variable of \code{refLinesData} with time for reference line(s)}

\item{refLinesLabelVar}{string, variable of \code{refLinesData} with label for reference line(s)}

\item{shiny}{logical, set to TRUE (FALSE by default) if the report is generated from a Shiny application.}

\item{verbose}{logical, if TRUE print messages during execution}

\item{nCores}{Integer containing the number of cores used for the computation
(1 by default). If more than 1, computation is parallelized, in this case
the package \code{parallel} is required.}
}
\value{
a list of \code{subjectProfilePlot} object, containing the combined
profile plots for each subject.
}
\description{
Combine subject profile plots.
}
\author{
Laure Cougnaud
}
