% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubjectProfileReport.R
\name{subjectProfileCombine}
\alias{subjectProfileCombine}
\title{Combine subject profile plots.}
\usage{
subjectProfileCombine(listPlots,
  timeLim = getXLimSubjectProfilePlots(listPlots), subjectVar = "USUBJID",
  refLines = NULL, refLinesData = NULL, refLinesTimeVar = NULL,
  refLinesLabelVar = NULL)
}
\arguments{
\item{listPlots}{list of \code{\link[ggplot2]{ggplot2}} objects}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} and maximum \code{timeEndVar}}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{refLines}{(optional) nested list with details for reference line(s).
Each sublist contains:
\itemize{
\item{(required) 'label': }{string with label for the reference line}
\item{(required) 'time': }{time(x) coordinate for the reference line,
'dotted' by default}
\item{(optional) 'color': }{color for the reference line,
'black' by default}
\item{(optional) 'linetype': }{linetype for the reference line}
}}

\item{refLinesData}{data.frame with data from which the reference line(s) should be extracted}

\item{refLinesTimeVar}{string, variable of \code{refLinesData} with time for reference line(s)}

\item{refLinesLabelVar}{string, variable of \code{refLinesData} with label for reference line(s)}
}
\value{
a list of \code{subjectProfilePlot} object, containing the combined
profile plots for each subject.
This is in essence only a \code{\link[ggplot2]{ggplot2}} objects,
(with the additional attribute 'nLinesPlot')
}
\description{
Combine subject profile plots.
}
\author{
Laure Cougnaud
}
