% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileEventPlot.R
\name{subjectProfileEventPlot}
\alias{subjectProfileEventPlot}
\title{Create plot of subject profiles for events}
\usage{
subjectProfileEventPlot(data, paramVar,
  paramLab = toString(getLabelVar(paramVar, labelVars = labelVars)),
  paramGroupVar = NULL, colorVar = NULL, colorLab = getLabelVar(colorVar,
  labelVars = labelVars), colorPalette = if (!is.null(colorVar))
  getPatientColorPalette(x = data[, colorVar]), shapeVar = colorVar,
  shapeLab = getLabelVar(shapeVar, labelVars = labelVars), shapePalette = if
  (!is.null(shapeVar)) getPatientShapePalette(x = data[, colorVar]), timeVar,
  subjectVar = "USUBJID", xLab = getLabelVar(timeVar, labelVars =
  labelVars), yLab = "", timeLim = NULL, title = paramLab,
  label = title, labelVars = NULL)
}
\arguments{
\item{data}{data.frame with data}

\item{paramVar}{string, variable of \code{data} with parameter (used in the y-axis)}

\item{paramLab}{string, label for \code{paramVar}}

\item{paramGroupVar}{(optional) string, variable of \code{data} with grouping.
If specified, the parameters will be grouped by this variable in the y-axis.}

\item{colorVar}{string, variable of \code{data} with color.}

\item{colorLab}{string, label for \code{colorVar}}

\item{colorPalette}{named vector with color for \code{colorVar}}

\item{shapeVar}{string, variable of \code{data} for shape of the points}

\item{shapeLab}{string, label for \code{shapeVar}}

\item{shapePalette}{named vector with shape for \code{shapeVar}}

\item{timeVar}{string, variable of \code{data} with time}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{xLab}{string, label for the x-axis}

\item{yLab}{string, label for the y-axis}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{startVar} and maximum \code{endVar}}

\item{title}{string, title for the plot}

\item{label}{string, label used in the plot.
This label is used to report the name of the panel 
in the text when the plots are combined (e.g. if the plot is empty).}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
list of \code{\link[ggplot2]{ggplot2} objects}, 
also of class \code{subjectProfileTextPlot}
}
\description{
Create plot of subject profiles for events
}
\author{
Laure Cougnaud
}
