% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileEventPlot.R
\name{subjectProfileEventPlot}
\alias{subjectProfileEventPlot}
\title{Create plot of subject profiles for events}
\usage{
subjectProfileEventPlot(
  data,
  paramVar,
  paramLab = toString(getLabelVar(paramVar, labelVars = labelVars)),
  paramVarSep = " - ",
  paramGroupVar = NULL,
  colorVar = NULL,
  colorLab = getLabelVar(colorVar, labelVars = labelVars),
  colorPalette = NULL,
  shapeVar = colorVar,
  shapeLab = getLabelVar(shapeVar, labelVars = labelVars),
  shapePalette = NULL,
  alpha = 1,
  timeVar,
  timeTrans = NULL,
  timeExpand = NULL,
  subjectVar = "USUBJID",
  subjectSubset = NULL,
  subsetData = NULL,
  subsetVar = NULL,
  subsetValue = NULL,
  xLab = getLabelVar(timeVar, labelVars = labelVars),
  yLab = "",
  timeLim = NULL,
  title = paramLab,
  label = title,
  labelVars = NULL,
  formatReport = subjectProfileReportFormat(),
  paging = TRUE
)
}
\arguments{
\item{data}{data.frame with data}

\item{paramVar}{string, variable of \code{data} with parameter (used in the y-axis)}

\item{paramLab}{string, label for \code{paramVar}}

\item{paramVarSep}{string with character(s) used to concatenate multiple 
\code{paramVar}, ' - ' by default.}

\item{paramGroupVar}{(optional) character vector with variable(s) of \code{data} with grouping.
If specified, the parameters will be grouped by this(these) variable(s) in the y-axis.}

\item{colorVar}{string, variable of \code{data} with color.}

\item{colorLab}{string, label for \code{colorVar}}

\item{colorPalette}{named vector with color for \code{colorVar}}

\item{shapeVar}{string, variable of \code{data} for shape of the points}

\item{shapeLab}{string, label for \code{shapeVar}}

\item{shapePalette}{named vector with shape for \code{shapeVar}}

\item{alpha}{Numeric with transparency, 1 by default.}

\item{timeVar}{string, variable of \code{data} with time}

\item{timeTrans}{ggplot2 transformation (see \code{\link[scales]{trans_new}}),
e.g. produced by the \code{\link{getTimeTrans}} function.}

\item{timeExpand}{Vector of range expansion constants for the time axis
(see \code{expand} parameter of the \code{\link[ggplot2]{scale_x_continuous}} function).}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{subjectSubset}{(optional) Character vector with subjects of interest 
(available in \code{subjectVar} of \code{data}), NULL by default.}

\item{subsetData}{Data.frame from which subset based on \code{subjectVar}/\code{subsetValue}
should be extracted. If not specified, \code{data} is used.
Records from \code{data} are then mapped by the \code{subjectVar} variable.}

\item{subsetVar}{variable of \code{data} used for subsetting}

\item{subsetValue}{character vector with value(s) of interest to consider for 
\code{subsetVar}}

\item{xLab}{string, label for the x-axis}

\item{yLab}{string, label for the y-axis}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} 
and maximum \code{timeEndVar} per subject.}

\item{title}{string, title for the plot}

\item{label}{string, label used in the plot.
This label is used to report the name of the panel 
in the text when the plots are combined (e.g. if the plot is empty).}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{formatReport}{list with parameters used to specify the format of the report,
e.g. output of the \code{\link{subjectProfileReportFormat}} function}

\item{paging}{Logical, if TRUE (by default), the automatic 
paging is enabled, otherwise only one page is used.}
}
\value{
list of (across subjects) of list (across modules) of \code{\link[ggplot2]{ggplot2} objects}, 
also of class \code{subjectProfileEventPlot}, with additional metaData attributes containing
'\code{label}', 'timeLim' and and 'timeTrans' (if specified).
}
\description{
Create plot of subject profiles for events
}
\author{
Laure Cougnaud
}
