% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileEventPlot.R
\name{subjectProfileEventPlot}
\alias{subjectProfileEventPlot}
\title{Create plot of subject profiles for events}
\usage{
subjectProfileEventPlot(
  data,
  paramVar,
  paramLab = toString(getLabelVar(paramVar, labelVars = labelVars)),
  paramVarSep = " - ",
  paramGroupVar = NULL,
  colorVar = NULL,
  colorLab = getLabelVar(colorVar, labelVars = labelVars),
  colorPalette = NULL,
  shapeVar = colorVar,
  shapeLab = getLabelVar(shapeVar, labelVars = labelVars),
  shapePalette = NULL,
  alpha = 1,
  timeVar,
  timeLab = getLabelVar(timeVar, labelVars = labelVars),
  timeTrans = NULL,
  timeExpand = NULL,
  subjectVar = "USUBJID",
  subjectSubset = NULL,
  subjectSample = NULL,
  seed = 123,
  subsetData = NULL,
  subsetVar = NULL,
  subsetValue = NULL,
  xLab = getLabelVar(timeVar, labelVars = labelVars),
  yLab = "",
  timeLim = NULL,
  title = toString(getLabelVar(paramVar, labelVars = labelVars, label = paramLab)),
  label = title,
  labelVars = NULL,
  formatReport = subjectProfileReportFormat(),
  paging = TRUE
)
}
\arguments{
\item{data}{Data.frame with data}

\item{paramVar}{Character vector with variable(s) of \code{data} 
with parameters. Variable content is displayed in the y-axis.}

\item{paramLab}{Named character vector, 
with label for the parameter variable(s) (\code{paramVar}).\cr
This is used to set the default title.}

\item{paramVarSep}{string with character(s) used to concatenate multiple 
\code{paramVar}, ' - ' by default.}

\item{paramGroupVar}{(optional) character vector with variable(s) of \code{data} with grouping.
If specified, the parameters will be grouped by this(these) variable(s) in the y-axis.}

\item{colorVar}{String, variable of \code{data} with color.
For the subject profile event plot, this variable is used
for the colors and the filling of the points.}

\item{colorLab}{String, label for \code{colorVar}.}

\item{colorPalette}{Named vector with color palette.
The variable should be named with the corresponding element
in \code{colorVar}.}

\item{shapeVar}{string, variable of \code{data} for shape of the points}

\item{shapeLab}{string, label for \code{shapeVar}}

\item{shapePalette}{named vector with shape for \code{shapeVar}}

\item{alpha}{Numeric with transparency, 1 by default.}

\item{timeVar}{string, variable of \code{data} with time}

\item{timeLab}{String, label for \code{timeVar}.}

\item{timeTrans}{ggplot2 transformation (see \code{\link[scales]{trans_new}}),
e.g. produced by the \code{\link{getTimeTrans}} function.}

\item{timeExpand}{Vector of range expansion constants for the time axis
(see \code{expand} parameter of the \code{\link[ggplot2]{scale_x_continuous}} function).}

\item{subjectVar}{String, variable of data (and subset data) with subject ID.}

\item{subjectSubset}{(optional) Character vector with subjects of interest 
(available in \code{subjectVar}), NULL by default.}

\item{subjectSample}{(optional) Integer of length 1
with number of random subject(s) that should be considered,
e.g. to check the created patient profiles for a subset of the data.
By default, all specified subjects are considered (set to NULL).}

\item{seed}{(optional) Integer of length 1 with seed used to select random subjects 
if \code{subjectSample} is specified (123 by default).}

\item{subsetData}{(optional) Data.frame with extra dataset to filter on.
This dataset is filtered, and only records from \code{data} 
with common subject IDs will be retained.\cr
If not specified, \code{data} is used.}

\item{subsetVar}{(optional) String with variable of subset data to filter on.
\code{subsetValue} should be specified too.\cr
If not specified, all records from the subset data are retained.}

\item{subsetValue}{(optional) Character vector with value(s) of interest to
retain in the filtered data.
These values should be available in \code{subsetVar}.\cr
Missing values in the subject variable are not retained 
in the filtered data.}

\item{xLab}{String, label for the x-axis.}

\item{yLab}{String, label for the y-axis.}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} 
and maximum \code{timeEndVar} per subject.}

\item{title}{string, title for the plot}

\item{label}{String, label for the visualization.
This label is used when the visualization is 
included in the final patient profile report,
in case this visualization is empty (no data available)
for a specific subject.}

\item{labelVars}{Named character vector with variable labels 
(names are the variable code)}

\item{formatReport}{list with parameters used to specify the format of the report,
e.g. output of the \code{\link{subjectProfileReportFormat}} function}

\item{paging}{Logical, if TRUE (by default), automatic 
paging is enabled, so patient profiles module too big to fit
in one page will span multiple pages.
Please note that the size of the graphic window 
(or report page) may need
to be re-sized in order that the plot fits.
\cr
If FALSE, the entire plot is included in one single page.}
}
\value{
list of (across subjects) of list (across modules) of \code{\link[ggplot2]{ggplot2} objects}, 
also of class \code{subjectProfileEventPlot}, with additional metaData attributes containing
'\code{label}', 'timeLim' and and 'timeTrans' (if specified).
}
\description{
Create plot of subject profiles for events
}
\seealso{
Other patient profiles plotting function: 
\code{\link{subjectProfileIntervalPlot}()},
\code{\link{subjectProfileLinePlot}()},
\code{\link{subjectProfileTextPlot}()}
}
\author{
Laure Cougnaud
}
\concept{patient profiles plotting function}
