% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileIntervalPlot.R
\name{subjectProfileIntervalPlot}
\alias{subjectProfileIntervalPlot}
\title{Create plot of subject profiles with segments for range of parameters}
\usage{
subjectProfileIntervalPlot(data, paramVar,
  paramLab = toString(getLabelVar(paramVar, labelVars = labelVars)),
  paramGroupVar = NULL, timeStartVar, timeEndVar, subjectVar = "USUBJID",
  timeLim = with(data, c(min(get(timeStartVar), na.rm = TRUE),
  max(get(timeEndVar), na.rm = TRUE))),
  rangeSimilarStartEnd = diff(timeLim)/1000,
  xLab = paste(getLabelVar(c(timeStartVar, timeEndVar), labelVars =
  labelVars), collapse = "/"), yLab = "", colorVar = NULL,
  colorLab = getLabelVar(colorVar, labelVars = labelVars),
  colorPalette = NULL, title = paramLab, label = title,
  labelVars = NULL)
}
\arguments{
\item{data}{data.frame with data}

\item{paramVar}{string, variable of \code{data} with parameter (used in the y-axis)}

\item{paramLab}{string, label for \code{paramVar}}

\item{timeStartVar}{string, variable of \code{data} with start of range}

\item{timeEndVar}{string, variable of \code{data} with end of range}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} and maximum \code{timeEndVar}}

\item{rangeSimilarStartEnd}{numeric, if a record has the same
\code{timeStartVar} and \code{timeEndVar}, what should be the range of the segment?
By default, a thousandth of the range of \code{timeLim}.}

\item{xLab}{string, label for the x-axis}

\item{yLab}{string, label for the y-axis}

\item{colorVar}{string, variable of \code{data} with color.}

\item{colorLab}{string, label for \code{colorVar}}

\item{colorPalette}{named vector with color for \code{colorVar}}

\item{title}{string, title for the plot}

\item{label}{string, label used in the plot.
This label is used to report the name of the panel 
in the text when the plots are combined (e.g. if the plot is empty).}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
list of \code{\link[ggplot2]{ggplot2} objects}, 
also of class \code{subjectProfileTextPlot}
}
\description{
Note that is no end date is available for a specific record,
the maximum of the time limits (\code{timeLim}) is used as end of the interval
and a triangle with right direction is added.
}
\author{
Laure Cougnaud
}
