% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileLinePlot.R
\name{subjectProfileLinePlot}
\alias{subjectProfileLinePlot}
\title{Create spaghetti plot for subject profiles}
\usage{
subjectProfileLinePlot(data, paramValueVar,
  paramLab = toString(getLabelVar(paramValueVar, labelVars = labelVars)),
  paramNameVar = NULL, paramValueRangeVar = NULL,
  colorValueRange = unname(glpgColor("extra")["lightGreen"]),
  colorVar = NULL, colorLab = getLabelVar(colorVar, labelVars =
  labelVars), colorPalette = NULL, shapeVar = colorVar,
  shapeLab = getLabelVar(shapeVar, labelVars = labelVars),
  shapePalette = NULL, paramGroupVar = NULL, timeVar,
  timeTrans = NULL, timeExpand = NULL, subjectVar = "USUBJID",
  subjectSubset = NULL, subsetData = NULL, subsetVar = NULL,
  subsetValue = NULL, xLab = getLabelVar(timeVar, labelVars =
  labelVars), yLab = "", timeLim = NULL, title = paramLab,
  label = title, labelVars = NULL,
  formatReport = subjectProfileReportFormat(), paging = TRUE,
  alpha = 1, yLimFrom = c("valueRange", "value"))
}
\arguments{
\item{data}{data.frame with data}

\item{paramValueVar}{string, variable of \code{data} with parameter value to represent}

\item{paramLab}{string, label for \code{paramVar}}

\item{paramNameVar}{string, variable of \code{data} with parameter name}

\item{paramValueRangeVar}{character vector of length 2 containing 
variables of \code{data} with minimum and maximum value 
for \code{paramValueVar}.
This range is represented as a ribbon in the plot background.
e.g. to represent the reference range of the variable.}

\item{colorValueRange}{String with color for the ribbon
represented by \code{paramValueRangeVar}.}

\item{colorVar}{string, variable of \code{data} with color, used for the points only.}

\item{colorLab}{string, label for \code{colorVar}}

\item{colorPalette}{named vector with color for \code{colorVar}}

\item{shapeVar}{string, variable of \code{data} with shape, used for the points only.}

\item{shapeLab}{string, label for \code{shapeVar}}

\item{shapePalette}{named vector with shape for \code{shapeVar}}

\item{paramGroupVar}{(optional) character vector with variable(s) of \code{data} with grouping.
If specified, the parameters will be grouped by this(these) variable(s) in the y-axis.}

\item{timeVar}{string, variable of \code{data} with time}

\item{timeTrans}{ggplot2 transformation (see \code{\link[scales]{trans_new}}),
e.g. produced by the \code{\link{getTimeTrans}} function.}

\item{timeExpand}{Vector of range expansion constants for the time axis
(see \code{expand} parameter of the \code{\link[ggplot2]{scale_x_continuous}} function).}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{subjectSubset}{(optional) Character vector with subjects of interest 
(available in \code{subjectVar} of \code{data}), NULL by default.}

\item{subsetData}{Data.frame from which subset based on \code{subjectVar}/\code{subsetValue}
should be extracted. If not specified, \code{data} is used.
Records from \code{data} are then mapped by the \code{subjectVar} variable.}

\item{subsetVar}{variable of \code{data} used for subsetting}

\item{subsetValue}{character vector with value(s) of interest to consider for 
\code{subsetVar}}

\item{xLab}{string, label for the x-axis}

\item{yLab}{string, label for the y-axis}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{timeStartVar} 
and maximum \code{timeEndVar} per subject.}

\item{title}{string, title for the plot}

\item{label}{string, label used in the plot.
This label is used to report the name of the panel 
in the text when the plots are combined (e.g. if the plot is empty).}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{formatReport}{list with parameters used to specify the format of the report,
e.g. output of the \code{\link{subjectProfileReportFormat}} function}

\item{paging}{Logical, if TRUE (by default), the automatic 
paging is enabled, otherwise only one page is used.}

\item{alpha}{Numeric with transparency, 1 by default.}

\item{yLimFrom}{String with specification on the limits of the y-axis, either:
\itemize{
\item{'valueRange' (by default): }{for each parameter, the y-axis range
spans the min/max specified via \code{paramValueRangeVar} (if specified),
the parameter value range otherwise}
\item{'value': }{the y-axis range is extracted from the min/max observed
value of each parameter available in \code{paramValueVar} (without considering the range
specified via \code{paramValueRangeVar}}
}}
}
\value{
List of (across subjects) of list (across modules) 
of \code{\link[ggplot2]{ggplot2} objects}, 
also of class \code{subjectProfileLinePlot}.
Each subject profile contains attributes: 'subjectID' and 'nLines' 
(estimated number of lines of space the plot will take).
The entire list also contains attributes: '\code{label}',
'timeLim' and 'timeTrans' (if specified).
}
\description{
Create spaghetti plot for subject profiles
}
\author{
Laure Cougnaud
}
