% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileRangePlot.R
\name{subjectProfileRangePlot}
\alias{subjectProfileRangePlot}
\title{Create plot of subject profiles with segments for range of parameters}
\usage{
subjectProfileRangePlot(data, paramVar, paramLab = getLabelVar(paramVar,
  labelVars = labelVars), startVar, endVar, subjectVar = "USUBJID",
  timeLim = with(data, c(min(get(startVar), na.rm = TRUE), max(get(startVar),
  na.rm = TRUE))), rangeSimilarStartEnd = diff(timeLim)/1000,
  xLab = paste(getLabelVar(c(startVar, endVar), labelVars = labelVars),
  collapse = "/"), yLab = "", colorVar = NULL,
  colorLab = getLabelVar(colorVar, labelVars = labelVars), title = paramLab,
  labelVars = NULL)
}
\arguments{
\item{data}{data.frame with data}

\item{paramVar}{string, variable of \code{data} with parameter (used in the y-axis)}

\item{paramLab}{string, label for \code{paramVar}}

\item{startVar}{string, variable of \code{data} with start of range}

\item{endVar}{string, variable of \code{data} with end of range}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{timeLim}{(optional) vector of length 2 with time limits (x-axis)
If not specified, extracted from the minimum \code{startVar} and maximum \code{endVar}}

\item{rangeSimilarStartEnd}{numeric, if a record has the same
\code{startVar} and \code{endVar}, what should be the range of the segment?
By default, a thousandth of the range of \code{timeLim}.}

\item{xLab}{string, label for the x-axis}

\item{yLab}{string, label for the y-axis}

\item{colorVar}{string, variable of \code{data} with color}

\item{colorLab}{string, label for \code{colorVar}}

\item{title}{string, title for the plot}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
list of \code{\link[ggplot2]{ggplot2} objects}, 
also of class \code{subjectProfileTextPlot}
}
\description{
Create plot of subject profiles with segments for range of parameters
}
\author{
Laure Cougnaud
}
