% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileTextPlot.R
\name{subjectProfileTextPlot}
\alias{subjectProfileTextPlot}
\title{Create plot of subject profiles for events.}
\usage{
subjectProfileTextPlot(data, paramValueVar, paramNameVar = NULL,
  subjectVar = "USUBJID", xLab = "", yLab = "",
  title = "Subject information", labelVars = NULL)
}
\arguments{
\item{data}{data.frame with data}

\item{paramValueVar}{string, variable of \code{data} containing the parameter
value to represent. This should be of length 1 if used in combination with
\code{paramNameVar}, otherwise multiple are possible.}

\item{paramNameVar}{(optional) string, variable of \code{data} with parameter name.
If specified, \code{paramValueVar} should be an unique variable.}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{xLab}{string, label for the x-axis}

\item{yLab}{string, label for the y-axis}

\item{title}{string, title for the plot}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
list of \code{\link[ggplot2]{ggplot2} objects}, also of class
\code{subjectProfileTextPlot}
}
\description{
There are two ways to specify the variables of interest to include:
\itemize{
\item{by specifying column(s) of interest containing parameter values, passed
to the \code{paramValueVar} parameter
}
\item{by specifying a combination of a variable containing the parameter name
(\code{paramNameVar}), coupled with a variable containing the 
parameter values (\code{paramValueVar})
}
}
}
\author{
Laure Cougnaud
}
