% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileTextPlot.R
\name{subjectProfileTextPlot}
\alias{subjectProfileTextPlot}
\title{Create plot of subject profiles for events.}
\usage{
subjectProfileTextPlot(
  data,
  paramValueVar,
  paramValueLab = getLabelVar(paramValueVar, labelVars = labelVars),
  paramNameVar = NULL,
  paramGroupVar = NULL,
  subsetData = NULL,
  subsetVar = NULL,
  subsetValue = NULL,
  subjectVar = "USUBJID",
  subjectSubset = NULL,
  subjectSample = NULL,
  seed = 123,
  xLab = "",
  yLab = "",
  title = "Subject information",
  label = title,
  labelVars = NULL,
  paramVarSep = " - ",
  formatReport = subjectProfileReportFormat(),
  paging = TRUE,
  table = FALSE,
  colWidth = NULL
)
}
\arguments{
\item{data}{data.frame with data}

\item{paramValueVar}{string, variable of \code{data}, either:
\itemize{
\item{if used in combination with \code{paramNameVar}: }{
\itemize{
\item{vector with names of variable(s) (multiple possible) 
of \code{data} with parameter value to represent.
}
\item{function taking \code{data} as input and 
returning a new variable (of length equal to number of rows in \code{data}
) with parameter value to represent}
}
}
\item{otherwise: }{vector with names of variable(s) (multiple possible) 
of \code{data} with parameter value to represent. 
If variables should be concatenated in the same line and not displayed as table: 
they should be specified separated by '|', e.g 'SEX|AGE'.
}
}}

\item{paramValueLab}{(optional) string with labels for \code{paramValueVar},
only used if the variabless are displayed as table.}

\item{paramNameVar}{(optional) string, variable of \code{data} with parameter name.
If specified, \code{paramValueVar} should be an unique variable.}

\item{paramGroupVar}{(optional) string, variable of \code{data} with grouping.
If specified, the parameters will be grouped by this variable in the y-axis, and
\code{paramValueVar} should be an unique variable.}

\item{subsetData}{Data.frame from which subset based on \code{subjectVar}/\code{subsetValue}
should be extracted. If not specified, \code{data} is used.
Records from \code{data} are then mapped by the \code{subjectVar} variable.}

\item{subsetVar}{variable of \code{data} used for subsetting}

\item{subsetValue}{character vector with value(s) of interest to consider for 
\code{subsetVar}}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{subjectSubset}{(optional) Character vector with subjects of interest 
(available in \code{subjectVar} of \code{data}), NULL by default.}

\item{subjectSample}{Integer with number of random subject(s) that should be considered,
e.g. to check the created patient profiles for a subset of the data.
By default, all specified subjects are considered (set to NULL).}

\item{seed}{Integer with seed used to select random subjects 
if \code{subjectSample} is specified.}

\item{xLab}{string, label for the x-axis}

\item{yLab}{string, label for the y-axis}

\item{title}{string, title for the plot}

\item{label}{string, label used in the plot.
This label is used to report the name of the panel 
in the text when the plots are combined (e.g. if the plot is empty).}

\item{labelVars}{Named character vector with variable labels 
(names are the variable code)}

\item{paramVarSep}{string with character(s) used to concatenate multiple 
\code{paramNameVar} or \code{paramValueVar}, ' - ' by default.}

\item{formatReport}{list with parameters used to specify the format of the report,
e.g. output of the \code{\link{subjectProfileReportFormat}} function}

\item{paging}{Logical, if TRUE (by default), the automatic 
paging is enabled, otherwise only one page is used.}

\item{table}{Logical, if TRUE (FALSE by default) the information
contained in the variables: \code{paramValueVar} is displayed as a table. 
Otherwise, the values of the different variables are concatenated in the same line.}

\item{colWidth}{Numeric vector of \code{length(paramValueVar)}
containing the approximate width of each parameter value column.
Only used if parameter values are displayed as a table (\code{table} is TRUE).
Note: columns can be slightly bigger if content larger than the specified width.
If not specified, optimized widths are determined.}
}
\value{
list of (across subjects) of list (across modules) of \code{\link[ggplot2]{ggplot2} objects}, 
also of class \code{subjectProfileTextPlot}, with additional metaData attributes containing
'\code{label}' and 'timeLim'.
}
\description{
There are two ways to specify the variables of interest to include:
\itemize{
\item{by specifying column(s) of interest containing parameter values, passed
to the \code{paramValueVar} parameter
}
\item{by specifying a combination of a variable containing the parameter name
(\code{paramNameVar}), coupled with a variable containing the 
parameter values (\code{paramValueVar})
}
}
}
\author{
Laure Cougnaud
}
