% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileTextPlot.R
\name{subjectProfileTextPlot}
\alias{subjectProfileTextPlot}
\title{Create plot of subject profiles for events.}
\usage{
subjectProfileTextPlot(data, paramValueVar, paramNameVar = NULL,
  paramGroupVar = NULL, subsetVar = NULL, subsetValue = NULL,
  subjectVar = "USUBJID", xLab = "", yLab = "",
  title = "Subject information", label = title, labelVars = NULL,
  paramVarSep = " - ", formatReport = subjectProfileReportFormat())
}
\arguments{
\item{data}{data.frame with data}

\item{paramValueVar}{string, variable of \code{data} containing the . 
Either:
\itemize{
\item{if used in combination with \code{paramNameVar}: }{
\itemize{
\item{vector with names of variable(s) (multiple possible) 
of \code{data} with parameter value to represent.
}
\item{function taking \code{data} as input and 
returning a new variable (of length equal to number of rows in \code{data}
) with parameter value to represent}
}
}
\item{otherwise: }{vector with names of variable(s) (multiple possible) 
of \code{data} with parameter value to represent. 
If variables should be concatenated in the same line, 
they should be specified separated by '|', e.g 'SEX|AGE'.
}
}}

\item{paramNameVar}{(optional) string, variable of \code{data} with parameter name.
If specified, \code{paramValueVar} should be an unique variable.}

\item{paramGroupVar}{(optional) string, variable of \code{data} with grouping.
If specified, the parameters will be grouped by this variable in the y-axis, and
\code{paramValueVar} should be an unique variable.}

\item{subsetVar}{variable of \code{data} used for subsetting}

\item{subsetValue}{character vector with value(s) of interest to consider for 
\code{subsetVar}}

\item{subjectVar}{string, variable of \code{data} with subject ID}

\item{xLab}{string, label for the x-axis}

\item{yLab}{string, label for the y-axis}

\item{title}{string, title for the plot}

\item{label}{string, label used in the plot.
This label is used to report the name of the panel 
in the text when the plots are combined (e.g. if the plot is empty).}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{paramVarSep}{string with character(s) used to concatenate multiple 
\code{paramNameVar} or \code{paramValueVar}, ' - ' by default.}

\item{formatReport}{list with parameters used to specify the format of the report,
e.g. output of the \code{\link{subjectProfileReportFormat}} function}
}
\value{
list of (across subjects) of list (across modules) of \code{\link[ggplot2]{ggplot2} objects}, 
also of class \code{subjectProfileTextPlot}, with additional metaData attributes containing
'label' and 'timeLim'.
}
\description{
There are two ways to specify the variables of interest to include:
\itemize{
\item{by specifying column(s) of interest containing parameter values, passed
to the \code{paramValueVar} parameter
}
\item{by specifying a combination of a variable containing the parameter name
(\code{paramNameVar}), coupled with a variable containing the 
parameter values (\code{paramValueVar})
}
}
}
\author{
Laure Cougnaud
}
